% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_service.R
\name{amplify}
\alias{amplify}
\title{AWS Amplify}
\usage{
amplify(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\value{
A client for the service. You can call the service's operations using
syntax like \code{svc$operation(...)}, where \code{svc} is the name you've assigned
to the client. The available operations are listed in the
Operations section.
}
\description{
Amplify enables developers to develop and deploy cloud-powered mobile
and web apps. The Amplify Console provides a continuous delivery and
hosting service for web applications. For more information, see the
\href{https://docs.aws.amazon.com/amplify/latest/userguide/welcome.html}{Amplify Console User Guide}.
The Amplify Framework is a comprehensive set of SDKs, libraries, tools,
and documentation for client app development. For more information, see
the \href{https://docs.amplify.aws/}{Amplify Framework.}
}
\section{Service syntax}{
\preformatted{svc <- amplify(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=amplify_create_app]{create_app} \tab Creates a new Amplify app\cr
\link[=amplify_create_backend_environment]{create_backend_environment} \tab Creates a new backend environment for an Amplify app\cr
\link[=amplify_create_branch]{create_branch} \tab Creates a new branch for an Amplify app\cr
\link[=amplify_create_deployment]{create_deployment} \tab Creates a deployment for a manually deployed Amplify app\cr
\link[=amplify_create_domain_association]{create_domain_association} \tab Creates a new domain association for an Amplify app\cr
\link[=amplify_create_webhook]{create_webhook} \tab Creates a new webhook on an Amplify app\cr
\link[=amplify_delete_app]{delete_app} \tab Deletes an existing Amplify app specified by an app ID\cr
\link[=amplify_delete_backend_environment]{delete_backend_environment} \tab Deletes a backend environment for an Amplify app\cr
\link[=amplify_delete_branch]{delete_branch} \tab Deletes a branch for an Amplify app\cr
\link[=amplify_delete_domain_association]{delete_domain_association} \tab Deletes a domain association for an Amplify app\cr
\link[=amplify_delete_job]{delete_job} \tab Deletes a job for a branch of an Amplify app\cr
\link[=amplify_delete_webhook]{delete_webhook} \tab Deletes a webhook\cr
\link[=amplify_generate_access_logs]{generate_access_logs} \tab Returns the website access logs for a specific time range using a presigned URL\cr
\link[=amplify_get_app]{get_app} \tab Returns an existing Amplify app by appID\cr
\link[=amplify_get_artifact_url]{get_artifact_url} \tab Returns the artifact info that corresponds to an artifact id\cr
\link[=amplify_get_backend_environment]{get_backend_environment} \tab Returns a backend environment for an Amplify app\cr
\link[=amplify_get_branch]{get_branch} \tab Returns a branch for an Amplify app\cr
\link[=amplify_get_domain_association]{get_domain_association} \tab Returns the domain information for an Amplify app\cr
\link[=amplify_get_job]{get_job} \tab Returns a job for a branch of an Amplify app\cr
\link[=amplify_get_webhook]{get_webhook} \tab Returns the webhook information that corresponds to a specified webhook ID\cr
\link[=amplify_list_apps]{list_apps} \tab Returns a list of the existing Amplify apps\cr
\link[=amplify_list_artifacts]{list_artifacts} \tab Returns a list of artifacts for a specified app, branch, and job\cr
\link[=amplify_list_backend_environments]{list_backend_environments} \tab Lists the backend environments for an Amplify app\cr
\link[=amplify_list_branches]{list_branches} \tab Lists the branches of an Amplify app\cr
\link[=amplify_list_domain_associations]{list_domain_associations} \tab Returns the domain associations for an Amplify app\cr
\link[=amplify_list_jobs]{list_jobs} \tab Lists the jobs for a branch of an Amplify app\cr
\link[=amplify_list_tags_for_resource]{list_tags_for_resource} \tab Returns a list of tags for a specified Amazon Resource Name (ARN)\cr
\link[=amplify_list_webhooks]{list_webhooks} \tab Returns a list of webhooks for an Amplify app\cr
\link[=amplify_start_deployment]{start_deployment} \tab Starts a deployment for a manually deployed app\cr
\link[=amplify_start_job]{start_job} \tab Starts a new job for a branch of an Amplify app\cr
\link[=amplify_stop_job]{stop_job} \tab Stops a job that is in progress for a branch of an Amplify app\cr
\link[=amplify_tag_resource]{tag_resource} \tab Tags the resource with a tag key and value\cr
\link[=amplify_untag_resource]{untag_resource} \tab Untags a resource with a specified Amazon Resource Name (ARN)\cr
\link[=amplify_update_app]{update_app} \tab Updates an existing Amplify app\cr
\link[=amplify_update_branch]{update_branch} \tab Updates a branch for an Amplify app\cr
\link[=amplify_update_domain_association]{update_domain_association} \tab Creates a new domain association for an Amplify app\cr
\link[=amplify_update_webhook]{update_webhook} \tab Updates a webhook
}
}

\examples{
\dontrun{
svc <- amplify()
svc$create_app(
  Foo = 123
)
}

}
