% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_create_app}
\alias{amplify_create_app}
\title{Creates a new Amplify app}
\usage{
amplify_create_app(name, description, repository, platform,
  iamServiceRoleArn, oauthToken, accessToken, environmentVariables,
  enableBranchAutoBuild, enableBranchAutoDeletion, enableBasicAuth,
  basicAuthCredentials, customRules, tags, buildSpec, customHeaders,
  enableAutoBranchCreation, autoBranchCreationPatterns,
  autoBranchCreationConfig)
}
\arguments{
\item{name}{[required] The name for an Amplify app.}

\item{description}{The description for an Amplify app.}

\item{repository}{The repository for an Amplify app.}

\item{platform}{The platform or framework for an Amplify app.}

\item{iamServiceRoleArn}{The AWS Identity and Access Management (IAM) service role for an Amplify
app.}

\item{oauthToken}{The OAuth token for a third-party source control system for an Amplify
app. The OAuth token is used to create a webhook and a read-only deploy
key. The OAuth token is not stored.}

\item{accessToken}{The personal access token for a third-party source control system for an
Amplify app. The personal access token is used to create a webhook and a
read-only deploy key. The token is not stored.}

\item{environmentVariables}{The environment variables map for an Amplify app.}

\item{enableBranchAutoBuild}{Enables the auto building of branches for an Amplify app.}

\item{enableBranchAutoDeletion}{Automatically disconnects a branch in the Amplify Console when you
delete a branch from your Git repository.}

\item{enableBasicAuth}{Enables basic authorization for an Amplify app. This will apply to all
branches that are part of this app.}

\item{basicAuthCredentials}{The credentials for basic authorization for an Amplify app.}

\item{customRules}{The custom rewrite and redirect rules for an Amplify app.}

\item{tags}{The tag for an Amplify app.}

\item{buildSpec}{The build specification (build spec) for an Amplify app.}

\item{customHeaders}{The custom HTTP headers for an Amplify app.}

\item{enableAutoBranchCreation}{Enables automated branch creation for an Amplify app.}

\item{autoBranchCreationPatterns}{The automated branch creation glob patterns for an Amplify app.}

\item{autoBranchCreationConfig}{The automated branch creation configuration for an Amplify app.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  app = list(
    appId = "string",
    appArn = "string",
    name = "string",
    tags = list(
      "string"
    ),
    description = "string",
    repository = "string",
    platform = "WEB",
    createTime = as.POSIXct(
      "2015-01-01"
    ),
    updateTime = as.POSIXct(
      "2015-01-01"
    ),
    iamServiceRoleArn = "string",
    environmentVariables = list(
      "string"
    ),
    defaultDomain = "string",
    enableBranchAutoBuild = TRUE|FALSE,
    enableBranchAutoDeletion = TRUE|FALSE,
    enableBasicAuth = TRUE|FALSE,
    basicAuthCredentials = "string",
    customRules = list(
      list(
        source = "string",
        target = "string",
        status = "string",
        condition = "string"
      )
    ),
    productionBranch = list(
      lastDeployTime = as.POSIXct(
        "2015-01-01"
      ),
      status = "string",
      thumbnailUrl = "string",
      branchName = "string"
    ),
    buildSpec = "string",
    customHeaders = "string",
    enableAutoBranchCreation = TRUE|FALSE,
    autoBranchCreationPatterns = list(
      "string"
    ),
    autoBranchCreationConfig = list(
      stage = "PRODUCTION"|"BETA"|"DEVELOPMENT"|"EXPERIMENTAL"|"PULL_REQUEST",
      framework = "string",
      enableAutoBuild = TRUE|FALSE,
      environmentVariables = list(
        "string"
      ),
      basicAuthCredentials = "string",
      enableBasicAuth = TRUE|FALSE,
      enablePerformanceMode = TRUE|FALSE,
      buildSpec = "string",
      enablePullRequestPreview = TRUE|FALSE,
      pullRequestEnvironmentName = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Creates a new Amplify app.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_app(
  name = "string",
  description = "string",
  repository = "string",
  platform = "WEB",
  iamServiceRoleArn = "string",
  oauthToken = "string",
  accessToken = "string",
  environmentVariables = list(
    "string"
  ),
  enableBranchAutoBuild = TRUE|FALSE,
  enableBranchAutoDeletion = TRUE|FALSE,
  enableBasicAuth = TRUE|FALSE,
  basicAuthCredentials = "string",
  customRules = list(
    list(
      source = "string",
      target = "string",
      status = "string",
      condition = "string"
    )
  ),
  tags = list(
    "string"
  ),
  buildSpec = "string",
  customHeaders = "string",
  enableAutoBranchCreation = TRUE|FALSE,
  autoBranchCreationPatterns = list(
    "string"
  ),
  autoBranchCreationConfig = list(
    stage = "PRODUCTION"|"BETA"|"DEVELOPMENT"|"EXPERIMENTAL"|"PULL_REQUEST",
    framework = "string",
    enableAutoBuild = TRUE|FALSE,
    environmentVariables = list(
      "string"
    ),
    basicAuthCredentials = "string",
    enableBasicAuth = TRUE|FALSE,
    enablePerformanceMode = TRUE|FALSE,
    buildSpec = "string",
    enablePullRequestPreview = TRUE|FALSE,
    pullRequestEnvironmentName = "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
