% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_create_deployment}
\alias{amplify_create_deployment}
\title{Create a deployment for manual deploy apps}
\usage{
amplify_create_deployment(appId, branchName, fileMap)
}
\arguments{
\item{appId}{[required] Unique Id for an Amplify App.}

\item{branchName}{[required] Name for the branch, for the Job.}

\item{fileMap}{Optional file map that contains file name as the key and file content
md5 hash as the value. If this argument is provided, the service will
generate different upload url per file. Otherwise, the service will only
generate a single upload url for the zipped files.}
}
\description{
Create a deployment for manual deploy apps. (Apps are not connected to
repository)
}
\section{Request syntax}{
\preformatted{svc$create_deployment(
  appId = "string",
  branchName = "string",
  fileMap = list(
    "string"
  )
)
}
}

\keyword{internal}
