% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_delete_app}
\alias{amplify_delete_app}
\title{Deletes an existing Amplify app specified by an app ID}
\usage{
amplify_delete_app(appId)
}
\arguments{
\item{appId}{[required] The unique ID for an Amplify app.}
}
\description{
Deletes an existing Amplify app specified by an app ID.
}
\section{Request syntax}{
\preformatted{svc$delete_app(
  appId = "string"
)
}
}

\keyword{internal}
