% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_delete_backend_environment}
\alias{amplify_delete_backend_environment}
\title{Deletes a backend environment for an Amplify app}
\usage{
amplify_delete_backend_environment(appId, environmentName)
}
\arguments{
\item{appId}{[required] The unique ID of an Amplify app.}

\item{environmentName}{[required] The name of a backend environment of an Amplify app.}
}
\value{
A list with the following syntax:\preformatted{list(
  backendEnvironment = list(
    backendEnvironmentArn = "string",
    environmentName = "string",
    stackName = "string",
    deploymentArtifacts = "string",
    createTime = as.POSIXct(
      "2015-01-01"
    ),
    updateTime = as.POSIXct(
      "2015-01-01"
    )
  )
)
}
}
\description{
Deletes a backend environment for an Amplify app.
}
\section{Request syntax}{
\preformatted{svc$delete_backend_environment(
  appId = "string",
  environmentName = "string"
)
}
}

\keyword{internal}
