% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_delete_branch}
\alias{amplify_delete_branch}
\title{Deletes a branch for an Amplify app}
\usage{
amplify_delete_branch(appId, branchName)
}
\arguments{
\item{appId}{[required] The unique ID for an Amplify app.}

\item{branchName}{[required] The name for the branch.}
}
\description{
Deletes a branch for an Amplify app.
}
\section{Request syntax}{
\preformatted{svc$delete_branch(
  appId = "string",
  branchName = "string"
)
}
}

\keyword{internal}
