% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_delete_domain_association}
\alias{amplify_delete_domain_association}
\title{Deletes a DomainAssociation}
\usage{
amplify_delete_domain_association(appId, domainName)
}
\arguments{
\item{appId}{[required] Unique Id for an Amplify App.}

\item{domainName}{[required] Name of the domain.}
}
\description{
Deletes a DomainAssociation.
}
\section{Request syntax}{
\preformatted{svc$delete_domain_association(
  appId = "string",
  domainName = "string"
)
}
}

\keyword{internal}
