% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_delete_webhook}
\alias{amplify_delete_webhook}
\title{Deletes a webhook}
\usage{
amplify_delete_webhook(webhookId)
}
\arguments{
\item{webhookId}{[required] The unique ID for a webhook.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  webhook = list(
    webhookArn = "string",
    webhookId = "string",
    webhookUrl = "string",
    branchName = "string",
    description = "string",
    createTime = as.POSIXct(
      "2015-01-01"
    ),
    updateTime = as.POSIXct(
      "2015-01-01"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Deletes a webhook.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_webhook(
  webhookId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
