% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_get_app}
\alias{amplify_get_app}
\title{Returns an existing Amplify app by appID}
\usage{
amplify_get_app(appId)
}
\arguments{
\item{appId}{[required] The unique ID for an Amplify app.}
}
\description{
Returns an existing Amplify app by appID.
}
\section{Request syntax}{
\preformatted{svc$get_app(
  appId = "string"
)
}
}

\keyword{internal}
