% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_get_artifact_url}
\alias{amplify_get_artifact_url}
\title{Returns the artifact info that corresponds to an artifact id}
\usage{
amplify_get_artifact_url(artifactId)
}
\arguments{
\item{artifactId}{[required] The unique ID for an artifact.}
}
\description{
Returns the artifact info that corresponds to an artifact id.
}
\section{Request syntax}{
\preformatted{svc$get_artifact_url(
  artifactId = "string"
)
}
}

\keyword{internal}
