% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_get_branch}
\alias{amplify_get_branch}
\title{Returns a branch for an Amplify app}
\usage{
amplify_get_branch(appId, branchName)
}
\arguments{
\item{appId}{[required] The unique ID for an Amplify app.}

\item{branchName}{[required] The name for the branch.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  branch = list(
    branchArn = "string",
    branchName = "string",
    description = "string",
    tags = list(
      "string"
    ),
    stage = "PRODUCTION"|"BETA"|"DEVELOPMENT"|"EXPERIMENTAL"|"PULL_REQUEST",
    displayName = "string",
    enableNotification = TRUE|FALSE,
    createTime = as.POSIXct(
      "2015-01-01"
    ),
    updateTime = as.POSIXct(
      "2015-01-01"
    ),
    environmentVariables = list(
      "string"
    ),
    enableAutoBuild = TRUE|FALSE,
    customDomains = list(
      "string"
    ),
    framework = "string",
    activeJobId = "string",
    totalNumberOfJobs = "string",
    enableBasicAuth = TRUE|FALSE,
    enablePerformanceMode = TRUE|FALSE,
    thumbnailUrl = "string",
    basicAuthCredentials = "string",
    buildSpec = "string",
    ttl = "string",
    associatedResources = list(
      "string"
    ),
    enablePullRequestPreview = TRUE|FALSE,
    pullRequestEnvironmentName = "string",
    destinationBranch = "string",
    sourceBranch = "string",
    backendEnvironmentArn = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns a branch for an Amplify app.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_branch(
  appId = "string",
  branchName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
