% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_get_domain_association}
\alias{amplify_get_domain_association}
\title{Returns the domain information for an Amplify app}
\usage{
amplify_get_domain_association(appId, domainName)
}
\arguments{
\item{appId}{[required] The unique id for an Amplify app.}

\item{domainName}{[required] The name of the domain.}
}
\description{
Returns the domain information for an Amplify app.
}
\section{Request syntax}{
\preformatted{svc$get_domain_association(
  appId = "string",
  domainName = "string"
)
}
}

\keyword{internal}
