% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_list_artifacts}
\alias{amplify_list_artifacts}
\title{List artifacts with an app, a branch, a job and an artifact type}
\usage{
amplify_list_artifacts(appId, branchName, jobId, nextToken, maxResults)
}
\arguments{
\item{appId}{[required] Unique Id for an Amplify App.}

\item{branchName}{[required] Name for a branch, part of an Amplify App.}

\item{jobId}{[required] Unique Id for an Job.}

\item{nextToken}{Pagination token. Set to null to start listing artifacts from start. If
non-null pagination token is returned in a result, then pass its value
in here to list more artifacts.}

\item{maxResults}{Maximum number of records to list in a single response.}
}
\description{
List artifacts with an app, a branch, a job and an artifact type.
}
\section{Request syntax}{
\preformatted{svc$list_artifacts(
  appId = "string",
  branchName = "string",
  jobId = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
