% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_list_backend_environments}
\alias{amplify_list_backend_environments}
\title{Lists backend environments for an Amplify App}
\usage{
amplify_list_backend_environments(appId, environmentName, nextToken,
  maxResults)
}
\arguments{
\item{appId}{[required] Unique Id for an amplify App.}

\item{environmentName}{Name of the backend environment}

\item{nextToken}{Pagination token. Set to null to start listing backen environments from
start. If a non-null pagination token is returned in a result, then pass
its value in here to list more backend environments.}

\item{maxResults}{Maximum number of records to list in a single response.}
}
\description{
Lists backend environments for an Amplify App.
}
\section{Request syntax}{
\preformatted{svc$list_backend_environments(
  appId = "string",
  environmentName = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
