% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_list_domain_associations}
\alias{amplify_list_domain_associations}
\title{Returns the domain associations for an Amplify app}
\usage{
amplify_list_domain_associations(appId, nextToken, maxResults)
}
\arguments{
\item{appId}{[required] The unique ID for an Amplify app.}

\item{nextToken}{A pagination token. Set to null to start listing apps from the start. If
non-null, a pagination token is returned in a result. Pass its value in
here to list more projects.}

\item{maxResults}{The maximum number of records to list in a single response.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  domainAssociations = list(
    list(
      domainAssociationArn = "string",
      domainName = "string",
      enableAutoSubDomain = TRUE|FALSE,
      autoSubDomainCreationPatterns = list(
        "string"
      ),
      autoSubDomainIAMRole = "string",
      domainStatus = "PENDING_VERIFICATION"|"IN_PROGRESS"|"AVAILABLE"|"PENDING_DEPLOYMENT"|"FAILED"|"CREATING"|"REQUESTING_CERTIFICATE"|"UPDATING",
      statusReason = "string",
      certificateVerificationDNSRecord = "string",
      subDomains = list(
        list(
          subDomainSetting = list(
            prefix = "string",
            branchName = "string"
          ),
          verified = TRUE|FALSE,
          dnsRecord = "string"
        )
      )
    )
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Returns the domain associations for an Amplify app.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_domain_associations(
  appId = "string",
  nextToken = "string",
  maxResults = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
