% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_list_jobs}
\alias{amplify_list_jobs}
\title{Lists the jobs for a branch of an Amplify app}
\usage{
amplify_list_jobs(appId, branchName, nextToken, maxResults)
}
\arguments{
\item{appId}{[required] The unique ID for an Amplify app.}

\item{branchName}{[required] The name for a branch.}

\item{nextToken}{A pagination token. Set to null to start listing steps from the start.
If a non-null pagination token is returned in a result, pass its value
in here to list more steps.}

\item{maxResults}{The maximum number of records to list in a single response.}
}
\value{
A list with the following syntax:\preformatted{list(
  jobSummaries = list(
    list(
      jobArn = "string",
      jobId = "string",
      commitId = "string",
      commitMessage = "string",
      commitTime = as.POSIXct(
        "2015-01-01"
      ),
      startTime = as.POSIXct(
        "2015-01-01"
      ),
      status = "PENDING"|"PROVISIONING"|"RUNNING"|"FAILED"|"SUCCEED"|"CANCELLING"|"CANCELLED",
      endTime = as.POSIXct(
        "2015-01-01"
      ),
      jobType = "RELEASE"|"RETRY"|"MANUAL"|"WEB_HOOK"
    )
  ),
  nextToken = "string"
)
}
}
\description{
Lists the jobs for a branch of an Amplify app.
}
\section{Request syntax}{
\preformatted{svc$list_jobs(
  appId = "string",
  branchName = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
