% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_list_tags_for_resource}
\alias{amplify_list_tags_for_resource}
\title{Returns a list of tags for a specified Amazon Resource Name (ARN)}
\usage{
amplify_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) to use to list tags.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  tags = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns a list of tags for a specified Amazon Resource Name (ARN).
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_tags_for_resource(
  resourceArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
