% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_start_deployment}
\alias{amplify_start_deployment}
\title{Starts a deployment for a manually deployed app}
\usage{
amplify_start_deployment(appId, branchName, jobId, sourceUrl)
}
\arguments{
\item{appId}{[required] The unique ID for an Amplify app.}

\item{branchName}{[required] The name for the branch, for the job.}

\item{jobId}{The job ID for this deployment, generated by the create deployment
request.}

\item{sourceUrl}{The source URL for this deployment, used when calling start deployment
without create deployment. The source URL can be any HTTP GET URL that
is publicly accessible and downloads a single .zip file.}
}
\description{
Starts a deployment for a manually deployed app. Manually deployed apps
are not connected to a repository.
}
\section{Request syntax}{
\preformatted{svc$start_deployment(
  appId = "string",
  branchName = "string",
  jobId = "string",
  sourceUrl = "string"
)
}
}

\keyword{internal}
