% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_start_job}
\alias{amplify_start_job}
\title{Starts a new job for a branch of an Amplify app}
\usage{
amplify_start_job(appId, branchName, jobId, jobType, jobReason,
  commitId, commitMessage, commitTime)
}
\arguments{
\item{appId}{[required] The unique ID for an Amplify app.}

\item{branchName}{[required] The branch name for the job.}

\item{jobId}{The unique ID for an existing job. This is required if the value of
\code{jobType} is \code{RETRY}.}

\item{jobType}{[required] Describes the type for the job. The job type \code{RELEASE} starts a new job
with the latest change from the specified branch. This value is
available only for apps that are connected to a repository. The job type
\code{RETRY} retries an existing job. If the job type value is \code{RETRY}, the
\code{jobId} is also required.}

\item{jobReason}{A descriptive reason for starting this job.}

\item{commitId}{The commit ID from a third-party repository provider for the job.}

\item{commitMessage}{The commit message from a third-party repository provider for the job.}

\item{commitTime}{The commit date and time for the job.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  jobSummary = list(
    jobArn = "string",
    jobId = "string",
    commitId = "string",
    commitMessage = "string",
    commitTime = as.POSIXct(
      "2015-01-01"
    ),
    startTime = as.POSIXct(
      "2015-01-01"
    ),
    status = "PENDING"|"PROVISIONING"|"RUNNING"|"FAILED"|"SUCCEED"|"CANCELLING"|"CANCELLED",
    endTime = as.POSIXct(
      "2015-01-01"
    ),
    jobType = "RELEASE"|"RETRY"|"MANUAL"|"WEB_HOOK"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Starts a new job for a branch of an Amplify app.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$start_job(
  appId = "string",
  branchName = "string",
  jobId = "string",
  jobType = "RELEASE"|"RETRY"|"MANUAL"|"WEB_HOOK",
  jobReason = "string",
  commitId = "string",
  commitMessage = "string",
  commitTime = as.POSIXct(
    "2015-01-01"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
