% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_stop_job}
\alias{amplify_stop_job}
\title{Stops a job that is in progress for a branch of an Amplify app}
\usage{
amplify_stop_job(appId, branchName, jobId)
}
\arguments{
\item{appId}{[required] The unique ID for an Amplify app.}

\item{branchName}{[required] The name for the branch, for the job.}

\item{jobId}{[required] The unique id for the job.}
}
\value{
A list with the following syntax:\preformatted{list(
  jobSummary = list(
    jobArn = "string",
    jobId = "string",
    commitId = "string",
    commitMessage = "string",
    commitTime = as.POSIXct(
      "2015-01-01"
    ),
    startTime = as.POSIXct(
      "2015-01-01"
    ),
    status = "PENDING"|"PROVISIONING"|"RUNNING"|"FAILED"|"SUCCEED"|"CANCELLING"|"CANCELLED",
    endTime = as.POSIXct(
      "2015-01-01"
    ),
    jobType = "RELEASE"|"RETRY"|"MANUAL"|"WEB_HOOK"
  )
)
}
}
\description{
Stops a job that is in progress for a branch of an Amplify app.
}
\section{Request syntax}{
\preformatted{svc$stop_job(
  appId = "string",
  branchName = "string",
  jobId = "string"
)
}
}

\keyword{internal}
