% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_tag_resource}
\alias{amplify_tag_resource}
\title{Tags the resource with a tag key and value}
\usage{
amplify_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) to use to tag a resource.}

\item{tags}{[required] The tags used to tag the resource.}
}
\description{
Tags the resource with a tag key and value.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  resourceArn = "string",
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
