% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_untag_resource}
\alias{amplify_untag_resource}
\title{Untags a resource with a specified Amazon Resource Name (ARN)}
\usage{
amplify_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) to use to untag a resource.}

\item{tagKeys}{[required] The tag keys to use to untag a resource.}
}
\description{
Untags a resource with a specified Amazon Resource Name (ARN).
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  resourceArn = "string",
  tagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
