% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_update_app}
\alias{amplify_update_app}
\title{Updates an existing Amplify app}
\usage{
amplify_update_app(appId, name, description, platform,
  iamServiceRoleArn, environmentVariables, enableBranchAutoBuild,
  enableBranchAutoDeletion, enableBasicAuth, basicAuthCredentials,
  customRules, buildSpec, customHeaders, enableAutoBranchCreation,
  autoBranchCreationPatterns, autoBranchCreationConfig, repository,
  oauthToken, accessToken)
}
\arguments{
\item{appId}{[required] The unique ID for an Amplify app.}

\item{name}{The name for an Amplify app.}

\item{description}{The description for an Amplify app.}

\item{platform}{The platform for an Amplify app.}

\item{iamServiceRoleArn}{The AWS Identity and Access Management (IAM) service role for an Amplify
app.}

\item{environmentVariables}{The environment variables for an Amplify app.}

\item{enableBranchAutoBuild}{Enables branch auto-building for an Amplify app.}

\item{enableBranchAutoDeletion}{Automatically disconnects a branch in the Amplify Console when you
delete a branch from your Git repository.}

\item{enableBasicAuth}{Enables basic authorization for an Amplify app.}

\item{basicAuthCredentials}{The basic authorization credentials for an Amplify app.}

\item{customRules}{The custom redirect and rewrite rules for an Amplify app.}

\item{buildSpec}{The build specification (build spec) for an Amplify app.}

\item{customHeaders}{The custom HTTP headers for an Amplify app.}

\item{enableAutoBranchCreation}{Enables automated branch creation for an Amplify app.}

\item{autoBranchCreationPatterns}{Describes the automated branch creation glob patterns for an Amplify
app.}

\item{autoBranchCreationConfig}{The automated branch creation configuration for an Amplify app.}

\item{repository}{The name of the repository for an Amplify app}

\item{oauthToken}{The OAuth token for a third-party source control system for an Amplify
app. The token is used to create a webhook and a read-only deploy key.
The OAuth token is not stored.}

\item{accessToken}{The personal access token for a third-party source control system for an
Amplify app. The token is used to create webhook and a read-only deploy
key. The token is not stored.}
}
\value{
A list with the following syntax:\preformatted{list(
  app = list(
    appId = "string",
    appArn = "string",
    name = "string",
    tags = list(
      "string"
    ),
    description = "string",
    repository = "string",
    platform = "WEB",
    createTime = as.POSIXct(
      "2015-01-01"
    ),
    updateTime = as.POSIXct(
      "2015-01-01"
    ),
    iamServiceRoleArn = "string",
    environmentVariables = list(
      "string"
    ),
    defaultDomain = "string",
    enableBranchAutoBuild = TRUE|FALSE,
    enableBranchAutoDeletion = TRUE|FALSE,
    enableBasicAuth = TRUE|FALSE,
    basicAuthCredentials = "string",
    customRules = list(
      list(
        source = "string",
        target = "string",
        status = "string",
        condition = "string"
      )
    ),
    productionBranch = list(
      lastDeployTime = as.POSIXct(
        "2015-01-01"
      ),
      status = "string",
      thumbnailUrl = "string",
      branchName = "string"
    ),
    buildSpec = "string",
    customHeaders = "string",
    enableAutoBranchCreation = TRUE|FALSE,
    autoBranchCreationPatterns = list(
      "string"
    ),
    autoBranchCreationConfig = list(
      stage = "PRODUCTION"|"BETA"|"DEVELOPMENT"|"EXPERIMENTAL"|"PULL_REQUEST",
      framework = "string",
      enableAutoBuild = TRUE|FALSE,
      environmentVariables = list(
        "string"
      ),
      basicAuthCredentials = "string",
      enableBasicAuth = TRUE|FALSE,
      enablePerformanceMode = TRUE|FALSE,
      buildSpec = "string",
      enablePullRequestPreview = TRUE|FALSE,
      pullRequestEnvironmentName = "string"
    )
  )
)
}
}
\description{
Updates an existing Amplify app.
}
\section{Request syntax}{
\preformatted{svc$update_app(
  appId = "string",
  name = "string",
  description = "string",
  platform = "WEB",
  iamServiceRoleArn = "string",
  environmentVariables = list(
    "string"
  ),
  enableBranchAutoBuild = TRUE|FALSE,
  enableBranchAutoDeletion = TRUE|FALSE,
  enableBasicAuth = TRUE|FALSE,
  basicAuthCredentials = "string",
  customRules = list(
    list(
      source = "string",
      target = "string",
      status = "string",
      condition = "string"
    )
  ),
  buildSpec = "string",
  customHeaders = "string",
  enableAutoBranchCreation = TRUE|FALSE,
  autoBranchCreationPatterns = list(
    "string"
  ),
  autoBranchCreationConfig = list(
    stage = "PRODUCTION"|"BETA"|"DEVELOPMENT"|"EXPERIMENTAL"|"PULL_REQUEST",
    framework = "string",
    enableAutoBuild = TRUE|FALSE,
    environmentVariables = list(
      "string"
    ),
    basicAuthCredentials = "string",
    enableBasicAuth = TRUE|FALSE,
    enablePerformanceMode = TRUE|FALSE,
    buildSpec = "string",
    enablePullRequestPreview = TRUE|FALSE,
    pullRequestEnvironmentName = "string"
  ),
  repository = "string",
  oauthToken = "string",
  accessToken = "string"
)
}
}

\keyword{internal}
