% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_update_webhook}
\alias{amplify_update_webhook}
\title{Updates a webhook}
\usage{
amplify_update_webhook(webhookId, branchName, description)
}
\arguments{
\item{webhookId}{[required] The unique ID for a webhook.}

\item{branchName}{The name for a branch that is part of an Amplify app.}

\item{description}{The description for a webhook.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  webhook = list(
    webhookArn = "string",
    webhookId = "string",
    webhookUrl = "string",
    branchName = "string",
    description = "string",
    createTime = as.POSIXct(
      "2015-01-01"
    ),
    updateTime = as.POSIXct(
      "2015-01-01"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Updates a webhook.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_webhook(
  webhookId = "string",
  branchName = "string",
  description = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
