% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsync_operations.R
\name{appsync_create_api_cache}
\alias{appsync_create_api_cache}
\title{Creates a cache for the GraphQL API}
\usage{
appsync_create_api_cache(apiId, ttl, transitEncryptionEnabled,
  atRestEncryptionEnabled, apiCachingBehavior, type)
}
\arguments{
\item{apiId}{[required] The GraphQL API Id.}

\item{ttl}{[required] TTL in seconds for cache entries.

Valid values are between 1 and 3600 seconds.}

\item{transitEncryptionEnabled}{Transit encryption flag when connecting to cache. This setting cannot be
updated after creation.}

\item{atRestEncryptionEnabled}{At rest encryption flag for cache. This setting cannot be updated after
creation.}

\item{apiCachingBehavior}{[required] Caching behavior.
\itemize{
\item \strong{FULL\\_REQUEST\\_CACHING}: All requests are fully cached.
\item \strong{PER\\_RESOLVER\\_CACHING}: Individual resolvers that you specify
are cached.
}}

\item{type}{[required] The cache instance type. Valid values are
\itemize{
\item \code{SMALL}
\item \code{MEDIUM}
\item \code{LARGE}
\item \code{XLARGE}
\item \code{LARGE_2X}
\item \code{LARGE_4X}
\item \code{LARGE_8X} (not available in all regions)
\item \code{LARGE_12X}
}

Historically, instance types were identified by an EC2-style value. As
of July 2020, this is deprecated, and the generic identifiers above
should be used.

The following legacy instance types are available, but their use is
discouraged:
\itemize{
\item \strong{T2\\_SMALL}: A t2.small instance type.
\item \strong{T2\\_MEDIUM}: A t2.medium instance type.
\item \strong{R4\\_LARGE}: A r4.large instance type.
\item \strong{R4\\_XLARGE}: A r4.xlarge instance type.
\item \strong{R4\\_2XLARGE}: A r4.2xlarge instance type.
\item \strong{R4\\_4XLARGE}: A r4.4xlarge instance type.
\item \strong{R4\\_8XLARGE}: A r4.8xlarge instance type.
}}
}
\description{
Creates a cache for the GraphQL API.
}
\section{Request syntax}{
\preformatted{svc$create_api_cache(
  apiId = "string",
  ttl = 123,
  transitEncryptionEnabled = TRUE|FALSE,
  atRestEncryptionEnabled = TRUE|FALSE,
  apiCachingBehavior = "FULL_REQUEST_CACHING"|"PER_RESOLVER_CACHING",
  type = "T2_SMALL"|"T2_MEDIUM"|"R4_LARGE"|"R4_XLARGE"|"R4_2XLARGE"|"R4_4XLARGE"|"R4_8XLARGE"|"SMALL"|"MEDIUM"|"LARGE"|"XLARGE"|"LARGE_2X"|"LARGE_4X"|"LARGE_8X"|"LARGE_12X"
)
}
}

\keyword{internal}
