% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsync_operations.R
\name{appsync_create_api_key}
\alias{appsync_create_api_key}
\title{Creates a unique key that you can distribute to clients who are
executing your API}
\usage{
appsync_create_api_key(apiId, description, expires)
}
\arguments{
\item{apiId}{[required] The ID for your GraphQL API.}

\item{description}{A description of the purpose of the API key.}

\item{expires}{The time from creation time after which the API key expires. The date is
represented as seconds since the epoch, rounded down to the nearest
hour. The default value for this parameter is 7 days from creation time.
For more information, see .}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  apiKey = list(
    id = "string",
    description = "string",
    expires = 123,
    deletes = 123
  )
)
}\if{html}{\out{</div>}}
}
\description{
Creates a unique key that you can distribute to clients who are
executing your API.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_api_key(
  apiId = "string",
  description = "string",
  expires = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
