% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsync_operations.R
\name{appsync_create_data_source}
\alias{appsync_create_data_source}
\title{Creates a DataSource object}
\usage{
appsync_create_data_source(apiId, name, description, type,
  serviceRoleArn, dynamodbConfig, lambdaConfig, elasticsearchConfig,
  httpConfig, relationalDatabaseConfig)
}
\arguments{
\item{apiId}{[required] The API ID for the GraphQL API for the \code{DataSource}.}

\item{name}{[required] A user-supplied name for the \code{DataSource}.}

\item{description}{A description of the \code{DataSource}.}

\item{type}{[required] The type of the \code{DataSource}.}

\item{serviceRoleArn}{The AWS IAM service role ARN for the data source. The system assumes
this role when accessing the data source.}

\item{dynamodbConfig}{Amazon DynamoDB settings.}

\item{lambdaConfig}{AWS Lambda settings.}

\item{elasticsearchConfig}{Amazon Elasticsearch Service settings.}

\item{httpConfig}{HTTP endpoint settings.}

\item{relationalDatabaseConfig}{Relational database settings.}
}
\value{
A list with the following syntax:\preformatted{list(
  dataSource = list(
    dataSourceArn = "string",
    name = "string",
    description = "string",
    type = "AWS_LAMBDA"|"AMAZON_DYNAMODB"|"AMAZON_ELASTICSEARCH"|"NONE"|"HTTP"|"RELATIONAL_DATABASE",
    serviceRoleArn = "string",
    dynamodbConfig = list(
      tableName = "string",
      awsRegion = "string",
      useCallerCredentials = TRUE|FALSE,
      deltaSyncConfig = list(
        baseTableTTL = 123,
        deltaSyncTableName = "string",
        deltaSyncTableTTL = 123
      ),
      versioned = TRUE|FALSE
    ),
    lambdaConfig = list(
      lambdaFunctionArn = "string"
    ),
    elasticsearchConfig = list(
      endpoint = "string",
      awsRegion = "string"
    ),
    httpConfig = list(
      endpoint = "string",
      authorizationConfig = list(
        authorizationType = "AWS_IAM",
        awsIamConfig = list(
          signingRegion = "string",
          signingServiceName = "string"
        )
      )
    ),
    relationalDatabaseConfig = list(
      relationalDatabaseSourceType = "RDS_HTTP_ENDPOINT",
      rdsHttpEndpointConfig = list(
        awsRegion = "string",
        dbClusterIdentifier = "string",
        databaseName = "string",
        schema = "string",
        awsSecretStoreArn = "string"
      )
    )
  )
)
}
}
\description{
Creates a \code{DataSource} object.
}
\section{Request syntax}{
\preformatted{svc$create_data_source(
  apiId = "string",
  name = "string",
  description = "string",
  type = "AWS_LAMBDA"|"AMAZON_DYNAMODB"|"AMAZON_ELASTICSEARCH"|"NONE"|"HTTP"|"RELATIONAL_DATABASE",
  serviceRoleArn = "string",
  dynamodbConfig = list(
    tableName = "string",
    awsRegion = "string",
    useCallerCredentials = TRUE|FALSE,
    deltaSyncConfig = list(
      baseTableTTL = 123,
      deltaSyncTableName = "string",
      deltaSyncTableTTL = 123
    ),
    versioned = TRUE|FALSE
  ),
  lambdaConfig = list(
    lambdaFunctionArn = "string"
  ),
  elasticsearchConfig = list(
    endpoint = "string",
    awsRegion = "string"
  ),
  httpConfig = list(
    endpoint = "string",
    authorizationConfig = list(
      authorizationType = "AWS_IAM",
      awsIamConfig = list(
        signingRegion = "string",
        signingServiceName = "string"
      )
    )
  ),
  relationalDatabaseConfig = list(
    relationalDatabaseSourceType = "RDS_HTTP_ENDPOINT",
    rdsHttpEndpointConfig = list(
      awsRegion = "string",
      dbClusterIdentifier = "string",
      databaseName = "string",
      schema = "string",
      awsSecretStoreArn = "string"
    )
  )
)
}
}

\keyword{internal}
