% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsync_operations.R
\name{appsync_create_function}
\alias{appsync_create_function}
\title{Creates a Function object}
\usage{
appsync_create_function(apiId, name, description, dataSourceName,
  requestMappingTemplate, responseMappingTemplate, functionVersion)
}
\arguments{
\item{apiId}{[required] The GraphQL API ID.}

\item{name}{[required] The \code{Function} name. The function name does not have to be unique.}

\item{description}{The \code{Function} description.}

\item{dataSourceName}{[required] The \code{Function} \code{DataSource} name.}

\item{requestMappingTemplate}{The \code{Function} request mapping template. Functions support only the
2018-05-29 version of the request mapping template.}

\item{responseMappingTemplate}{The \code{Function} response mapping template.}

\item{functionVersion}{[required] The \code{version} of the request mapping template. Currently the supported
value is 2018-05-29.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  functionConfiguration = list(
    functionId = "string",
    functionArn = "string",
    name = "string",
    description = "string",
    dataSourceName = "string",
    requestMappingTemplate = "string",
    responseMappingTemplate = "string",
    functionVersion = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Creates a \code{Function} object.

A function is a reusable entity. Multiple functions can be used to
compose the resolver logic.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_function(
  apiId = "string",
  name = "string",
  description = "string",
  dataSourceName = "string",
  requestMappingTemplate = "string",
  responseMappingTemplate = "string",
  functionVersion = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
