% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsync_operations.R
\name{appsync_create_type}
\alias{appsync_create_type}
\title{Creates a Type object}
\usage{
appsync_create_type(apiId, definition, format)
}
\arguments{
\item{apiId}{[required] The API ID.}

\item{definition}{[required] The type definition, in GraphQL Schema Definition Language (SDL) format.

For more information, see the \href{http://graphql.org/learn/schema/}{GraphQL SDL documentation}.}

\item{format}{[required] The type format: SDL or JSON.}
}
\description{
Creates a \code{Type} object.
}
\section{Request syntax}{
\preformatted{svc$create_type(
  apiId = "string",
  definition = "string",
  format = "SDL"|"JSON"
)
}
}

\keyword{internal}
