% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsync_operations.R
\name{appsync_delete_api_cache}
\alias{appsync_delete_api_cache}
\title{Deletes an ApiCache object}
\usage{
appsync_delete_api_cache(apiId)
}
\arguments{
\item{apiId}{[required] The API ID.}
}
\value{
An empty list.
}
\description{
Deletes an \code{ApiCache} object.
}
\section{Request syntax}{
\preformatted{svc$delete_api_cache(
  apiId = "string"
)
}
}

\keyword{internal}
