% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsync_operations.R
\name{appsync_delete_api_key}
\alias{appsync_delete_api_key}
\title{Deletes an API key}
\usage{
appsync_delete_api_key(apiId, id)
}
\arguments{
\item{apiId}{[required] The API ID.}

\item{id}{[required] The ID for the API key.}
}
\description{
Deletes an API key.
}
\section{Request syntax}{
\preformatted{svc$delete_api_key(
  apiId = "string",
  id = "string"
)
}
}

\keyword{internal}
