% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsync_operations.R
\name{appsync_delete_resolver}
\alias{appsync_delete_resolver}
\title{Deletes a Resolver object}
\usage{
appsync_delete_resolver(apiId, typeName, fieldName)
}
\arguments{
\item{apiId}{[required] The API ID.}

\item{typeName}{[required] The name of the resolver type.}

\item{fieldName}{[required] The resolver field name.}
}
\value{
An empty list.
}
\description{
Deletes a \code{Resolver} object.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_resolver(
  apiId = "string",
  typeName = "string",
  fieldName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
