% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsync_operations.R
\name{appsync_delete_type}
\alias{appsync_delete_type}
\title{Deletes a Type object}
\usage{
appsync_delete_type(apiId, typeName)
}
\arguments{
\item{apiId}{[required] The API ID.}

\item{typeName}{[required] The type name.}
}
\value{
An empty list.
}
\description{
Deletes a \code{Type} object.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_type(
  apiId = "string",
  typeName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
