% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsync_operations.R
\name{appsync_flush_api_cache}
\alias{appsync_flush_api_cache}
\title{Flushes an ApiCache object}
\usage{
appsync_flush_api_cache(apiId)
}
\arguments{
\item{apiId}{[required] The API ID.}
}
\value{
An empty list.
}
\description{
Flushes an \code{ApiCache} object.
}
\section{Request syntax}{
\preformatted{svc$flush_api_cache(
  apiId = "string"
)
}
}

\keyword{internal}
