% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsync_operations.R
\name{appsync_get_api_cache}
\alias{appsync_get_api_cache}
\title{Retrieves an ApiCache object}
\usage{
appsync_get_api_cache(apiId)
}
\arguments{
\item{apiId}{[required] The API ID.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  apiCache = list(
    ttl = 123,
    apiCachingBehavior = "FULL_REQUEST_CACHING"|"PER_RESOLVER_CACHING",
    transitEncryptionEnabled = TRUE|FALSE,
    atRestEncryptionEnabled = TRUE|FALSE,
    type = "T2_SMALL"|"T2_MEDIUM"|"R4_LARGE"|"R4_XLARGE"|"R4_2XLARGE"|"R4_4XLARGE"|"R4_8XLARGE"|"SMALL"|"MEDIUM"|"LARGE"|"XLARGE"|"LARGE_2X"|"LARGE_4X"|"LARGE_8X"|"LARGE_12X",
    status = "AVAILABLE"|"CREATING"|"DELETING"|"MODIFYING"|"FAILED"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves an \code{ApiCache} object.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_api_cache(
  apiId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
