% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsync_operations.R
\name{appsync_get_data_source}
\alias{appsync_get_data_source}
\title{Retrieves a DataSource object}
\usage{
appsync_get_data_source(apiId, name)
}
\arguments{
\item{apiId}{[required] The API ID.}

\item{name}{[required] The name of the data source.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  dataSource = list(
    dataSourceArn = "string",
    name = "string",
    description = "string",
    type = "AWS_LAMBDA"|"AMAZON_DYNAMODB"|"AMAZON_ELASTICSEARCH"|"NONE"|"HTTP"|"RELATIONAL_DATABASE",
    serviceRoleArn = "string",
    dynamodbConfig = list(
      tableName = "string",
      awsRegion = "string",
      useCallerCredentials = TRUE|FALSE,
      deltaSyncConfig = list(
        baseTableTTL = 123,
        deltaSyncTableName = "string",
        deltaSyncTableTTL = 123
      ),
      versioned = TRUE|FALSE
    ),
    lambdaConfig = list(
      lambdaFunctionArn = "string"
    ),
    elasticsearchConfig = list(
      endpoint = "string",
      awsRegion = "string"
    ),
    httpConfig = list(
      endpoint = "string",
      authorizationConfig = list(
        authorizationType = "AWS_IAM",
        awsIamConfig = list(
          signingRegion = "string",
          signingServiceName = "string"
        )
      )
    ),
    relationalDatabaseConfig = list(
      relationalDatabaseSourceType = "RDS_HTTP_ENDPOINT",
      rdsHttpEndpointConfig = list(
        awsRegion = "string",
        dbClusterIdentifier = "string",
        databaseName = "string",
        schema = "string",
        awsSecretStoreArn = "string"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves a \code{DataSource} object.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_data_source(
  apiId = "string",
  name = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
