% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsync_operations.R
\name{appsync_get_introspection_schema}
\alias{appsync_get_introspection_schema}
\title{Retrieves the introspection schema for a GraphQL API}
\usage{
appsync_get_introspection_schema(apiId, format, includeDirectives)
}
\arguments{
\item{apiId}{[required] The API ID.}

\item{format}{[required] The schema format: SDL or JSON.}

\item{includeDirectives}{A flag that specifies whether the schema introspection should contain
directives.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  schema = raw
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves the introspection schema for a GraphQL API.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_introspection_schema(
  apiId = "string",
  format = "SDL"|"JSON",
  includeDirectives = TRUE|FALSE
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
