% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsync_operations.R
\name{appsync_get_resolver}
\alias{appsync_get_resolver}
\title{Retrieves a Resolver object}
\usage{
appsync_get_resolver(apiId, typeName, fieldName)
}
\arguments{
\item{apiId}{[required] The API ID.}

\item{typeName}{[required] The resolver type name.}

\item{fieldName}{[required] The resolver field name.}
}
\value{
A list with the following syntax:\preformatted{list(
  resolver = list(
    typeName = "string",
    fieldName = "string",
    dataSourceName = "string",
    resolverArn = "string",
    requestMappingTemplate = "string",
    responseMappingTemplate = "string",
    kind = "UNIT"|"PIPELINE",
    pipelineConfig = list(
      functions = list(
        "string"
      )
    ),
    syncConfig = list(
      conflictHandler = "OPTIMISTIC_CONCURRENCY"|"LAMBDA"|"AUTOMERGE"|"NONE",
      conflictDetection = "VERSION"|"NONE",
      lambdaConflictHandlerConfig = list(
        lambdaConflictHandlerArn = "string"
      )
    ),
    cachingConfig = list(
      ttl = 123,
      cachingKeys = list(
        "string"
      )
    )
  )
)
}
}
\description{
Retrieves a \code{Resolver} object.
}
\section{Request syntax}{
\preformatted{svc$get_resolver(
  apiId = "string",
  typeName = "string",
  fieldName = "string"
)
}
}

\keyword{internal}
