% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsync_operations.R
\name{appsync_get_schema_creation_status}
\alias{appsync_get_schema_creation_status}
\title{Retrieves the current status of a schema creation operation}
\usage{
appsync_get_schema_creation_status(apiId)
}
\arguments{
\item{apiId}{[required] The API ID.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  status = "PROCESSING"|"ACTIVE"|"DELETING"|"FAILED"|"SUCCESS"|"NOT_APPLICABLE",
  details = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves the current status of a schema creation operation.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_schema_creation_status(
  apiId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
