% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsync_operations.R
\name{appsync_list_api_keys}
\alias{appsync_list_api_keys}
\title{Lists the API keys for a given API}
\usage{
appsync_list_api_keys(apiId, nextToken, maxResults)
}
\arguments{
\item{apiId}{[required] The API ID.}

\item{nextToken}{An identifier that was returned from the previous call to this
operation, which can be used to return the next set of items in the
list.}

\item{maxResults}{The maximum number of results you want the request to return.}
}
\value{
A list with the following syntax:\preformatted{list(
  apiKeys = list(
    list(
      id = "string",
      description = "string",
      expires = 123,
      deletes = 123
    )
  ),
  nextToken = "string"
)
}
}
\description{
Lists the API keys for a given API.

API keys are deleted automatically 60 days after they expire. However,
they may still be included in the response until they have actually been
deleted. You can safely call \code{\link[=appsync_delete_api_key]{delete_api_key}}
to manually delete a key before it's automatically deleted.
}
\section{Request syntax}{
\preformatted{svc$list_api_keys(
  apiId = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
