% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsync_operations.R
\name{appsync_list_data_sources}
\alias{appsync_list_data_sources}
\title{Lists the data sources for a given API}
\usage{
appsync_list_data_sources(apiId, nextToken, maxResults)
}
\arguments{
\item{apiId}{[required] The API ID.}

\item{nextToken}{An identifier that was returned from the previous call to this
operation, which can be used to return the next set of items in the
list.}

\item{maxResults}{The maximum number of results you want the request to return.}
}
\value{
A list with the following syntax:\preformatted{list(
  dataSources = list(
    list(
      dataSourceArn = "string",
      name = "string",
      description = "string",
      type = "AWS_LAMBDA"|"AMAZON_DYNAMODB"|"AMAZON_ELASTICSEARCH"|"NONE"|"HTTP"|"RELATIONAL_DATABASE",
      serviceRoleArn = "string",
      dynamodbConfig = list(
        tableName = "string",
        awsRegion = "string",
        useCallerCredentials = TRUE|FALSE,
        deltaSyncConfig = list(
          baseTableTTL = 123,
          deltaSyncTableName = "string",
          deltaSyncTableTTL = 123
        ),
        versioned = TRUE|FALSE
      ),
      lambdaConfig = list(
        lambdaFunctionArn = "string"
      ),
      elasticsearchConfig = list(
        endpoint = "string",
        awsRegion = "string"
      ),
      httpConfig = list(
        endpoint = "string",
        authorizationConfig = list(
          authorizationType = "AWS_IAM",
          awsIamConfig = list(
            signingRegion = "string",
            signingServiceName = "string"
          )
        )
      ),
      relationalDatabaseConfig = list(
        relationalDatabaseSourceType = "RDS_HTTP_ENDPOINT",
        rdsHttpEndpointConfig = list(
          awsRegion = "string",
          dbClusterIdentifier = "string",
          databaseName = "string",
          schema = "string",
          awsSecretStoreArn = "string"
        )
      )
    )
  ),
  nextToken = "string"
)
}
}
\description{
Lists the data sources for a given API.
}
\section{Request syntax}{
\preformatted{svc$list_data_sources(
  apiId = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
