% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsync_operations.R
\name{appsync_list_functions}
\alias{appsync_list_functions}
\title{List multiple functions}
\usage{
appsync_list_functions(apiId, nextToken, maxResults)
}
\arguments{
\item{apiId}{[required] The GraphQL API ID.}

\item{nextToken}{An identifier that was returned from the previous call to this
operation, which can be used to return the next set of items in the
list.}

\item{maxResults}{The maximum number of results you want the request to return.}
}
\value{
A list with the following syntax:\preformatted{list(
  functions = list(
    list(
      functionId = "string",
      functionArn = "string",
      name = "string",
      description = "string",
      dataSourceName = "string",
      requestMappingTemplate = "string",
      responseMappingTemplate = "string",
      functionVersion = "string",
      syncConfig = list(
        conflictHandler = "OPTIMISTIC_CONCURRENCY"|"LAMBDA"|"AUTOMERGE"|"NONE",
        conflictDetection = "VERSION"|"NONE",
        lambdaConflictHandlerConfig = list(
          lambdaConflictHandlerArn = "string"
        )
      )
    )
  ),
  nextToken = "string"
)
}
}
\description{
List multiple functions.
}
\section{Request syntax}{
\preformatted{svc$list_functions(
  apiId = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
