% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsync_operations.R
\name{appsync_list_graphql_apis}
\alias{appsync_list_graphql_apis}
\title{Lists your GraphQL APIs}
\usage{
appsync_list_graphql_apis(nextToken, maxResults)
}
\arguments{
\item{nextToken}{An identifier that was returned from the previous call to this
operation, which can be used to return the next set of items in the
list.}

\item{maxResults}{The maximum number of results you want the request to return.}
}
\value{
A list with the following syntax:\preformatted{list(
  graphqlApis = list(
    list(
      name = "string",
      apiId = "string",
      authenticationType = "API_KEY"|"AWS_IAM"|"AMAZON_COGNITO_USER_POOLS"|"OPENID_CONNECT",
      logConfig = list(
        fieldLogLevel = "NONE"|"ERROR"|"ALL",
        cloudWatchLogsRoleArn = "string",
        excludeVerboseContent = TRUE|FALSE
      ),
      userPoolConfig = list(
        userPoolId = "string",
        awsRegion = "string",
        defaultAction = "ALLOW"|"DENY",
        appIdClientRegex = "string"
      ),
      openIDConnectConfig = list(
        issuer = "string",
        clientId = "string",
        iatTTL = 123,
        authTTL = 123
      ),
      arn = "string",
      uris = list(
        "string"
      ),
      tags = list(
        "string"
      ),
      additionalAuthenticationProviders = list(
        list(
          authenticationType = "API_KEY"|"AWS_IAM"|"AMAZON_COGNITO_USER_POOLS"|"OPENID_CONNECT",
          openIDConnectConfig = list(
            issuer = "string",
            clientId = "string",
            iatTTL = 123,
            authTTL = 123
          ),
          userPoolConfig = list(
            userPoolId = "string",
            awsRegion = "string",
            appIdClientRegex = "string"
          )
        )
      ),
      xrayEnabled = TRUE|FALSE,
      wafWebAclArn = "string"
    )
  ),
  nextToken = "string"
)
}
}
\description{
Lists your GraphQL APIs.
}
\section{Request syntax}{
\preformatted{svc$list_graphql_apis(
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
