% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsync_operations.R
\name{appsync_list_resolvers_by_function}
\alias{appsync_list_resolvers_by_function}
\title{List the resolvers that are associated with a specific function}
\usage{
appsync_list_resolvers_by_function(apiId, functionId, nextToken,
  maxResults)
}
\arguments{
\item{apiId}{[required] The API ID.}

\item{functionId}{[required] The Function ID.}

\item{nextToken}{An identifier that was returned from the previous call to this
operation, which you can use to return the next set of items in the
list.}

\item{maxResults}{The maximum number of results you want the request to return.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  resolvers = list(
    list(
      typeName = "string",
      fieldName = "string",
      dataSourceName = "string",
      resolverArn = "string",
      requestMappingTemplate = "string",
      responseMappingTemplate = "string",
      kind = "UNIT"|"PIPELINE",
      pipelineConfig = list(
        functions = list(
          "string"
        )
      ),
      syncConfig = list(
        conflictHandler = "OPTIMISTIC_CONCURRENCY"|"LAMBDA"|"AUTOMERGE"|"NONE",
        conflictDetection = "VERSION"|"NONE",
        lambdaConflictHandlerConfig = list(
          lambdaConflictHandlerArn = "string"
        )
      ),
      cachingConfig = list(
        ttl = 123,
        cachingKeys = list(
          "string"
        )
      )
    )
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
List the resolvers that are associated with a specific function.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_resolvers_by_function(
  apiId = "string",
  functionId = "string",
  nextToken = "string",
  maxResults = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
