% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsync_operations.R
\name{appsync_list_types}
\alias{appsync_list_types}
\title{Lists the types for a given API}
\usage{
appsync_list_types(apiId, format, nextToken, maxResults)
}
\arguments{
\item{apiId}{[required] The API ID.}

\item{format}{[required] The type format: SDL or JSON.}

\item{nextToken}{An identifier that was returned from the previous call to this
operation, which can be used to return the next set of items in the
list.}

\item{maxResults}{The maximum number of results you want the request to return.}
}
\value{
A list with the following syntax:\preformatted{list(
  types = list(
    list(
      name = "string",
      description = "string",
      arn = "string",
      definition = "string",
      format = "SDL"|"JSON"
    )
  ),
  nextToken = "string"
)
}
}
\description{
Lists the types for a given API.
}
\section{Request syntax}{
\preformatted{svc$list_types(
  apiId = "string",
  format = "SDL"|"JSON",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
