% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsync_operations.R
\name{appsync_start_schema_creation}
\alias{appsync_start_schema_creation}
\title{Adds a new schema to your GraphQL API}
\usage{
appsync_start_schema_creation(apiId, definition)
}
\arguments{
\item{apiId}{[required] The API ID.}

\item{definition}{[required] The schema definition, in GraphQL schema language format.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  status = "PROCESSING"|"ACTIVE"|"DELETING"|"FAILED"|"SUCCESS"|"NOT_APPLICABLE"
)
}\if{html}{\out{</div>}}
}
\description{
Adds a new schema to your GraphQL API.

This operation is asynchronous. Use to determine when it has completed.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$start_schema_creation(
  apiId = "string",
  definition = raw
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
