% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsync_operations.R
\name{appsync_tag_resource}
\alias{appsync_tag_resource}
\title{Tags a resource with user-supplied tags}
\usage{
appsync_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The \code{GraphqlApi} ARN.}

\item{tags}{[required] A \code{TagMap} object.}
}
\value{
An empty list.
}
\description{
Tags a resource with user-supplied tags.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  resourceArn = "string",
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
