% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsync_operations.R
\name{appsync_untag_resource}
\alias{appsync_untag_resource}
\title{Untags a resource}
\usage{
appsync_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The \code{GraphqlApi} ARN.}

\item{tagKeys}{[required] A list of \code{TagKey} objects.}
}
\description{
Untags a resource.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  resourceArn = "string",
  tagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
