% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsync_operations.R
\name{appsync_update_graphql_api}
\alias{appsync_update_graphql_api}
\title{Updates a GraphqlApi object}
\usage{
appsync_update_graphql_api(apiId, name, logConfig, authenticationType,
  userPoolConfig, openIDConnectConfig, additionalAuthenticationProviders)
}
\arguments{
\item{apiId}{[required] The API ID.}

\item{name}{[required] The new name for the \code{GraphqlApi} object.}

\item{logConfig}{The Amazon CloudWatch Logs configuration for the \code{GraphqlApi} object.}

\item{authenticationType}{The new authentication type for the \code{GraphqlApi} object.}

\item{userPoolConfig}{The new Amazon Cognito user pool configuration for the \code{GraphqlApi}
object.}

\item{openIDConnectConfig}{The OpenID Connect configuration for the \code{GraphqlApi} object.}

\item{additionalAuthenticationProviders}{A list of additional authentication providers for the \code{GraphqlApi} API.}
}
\description{
Updates a \code{GraphqlApi} object.
}
\section{Request syntax}{
\preformatted{svc$update_graphql_api(
  apiId = "string",
  name = "string",
  logConfig = list(
    fieldLogLevel = "NONE"|"ERROR"|"ALL",
    cloudWatchLogsRoleArn = "string"
  ),
  authenticationType = "API_KEY"|"AWS_IAM"|"AMAZON_COGNITO_USER_POOLS"|"OPENID_CONNECT",
  userPoolConfig = list(
    userPoolId = "string",
    awsRegion = "string",
    defaultAction = "ALLOW"|"DENY",
    appIdClientRegex = "string"
  ),
  openIDConnectConfig = list(
    issuer = "string",
    clientId = "string",
    iatTTL = 123,
    authTTL = 123
  ),
  additionalAuthenticationProviders = list(
    list(
      authenticationType = "API_KEY"|"AWS_IAM"|"AMAZON_COGNITO_USER_POOLS"|"OPENID_CONNECT",
      openIDConnectConfig = list(
        issuer = "string",
        clientId = "string",
        iatTTL = 123,
        authTTL = 123
      ),
      userPoolConfig = list(
        userPoolId = "string",
        awsRegion = "string",
        appIdClientRegex = "string"
      )
    )
  )
)
}
}

\keyword{internal}
