% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mobile_operations.R
\name{delete_project}
\alias{delete_project}
\title{Delets a project in AWS Mobile Hub}
\usage{
delete_project(
  projectId = "string"
)
}
\arguments{
\item{projectId}{[required] Unique project identifier.}
}
\description{
Delets a project in AWS Mobile Hub.
}
\examples{

}
