% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_create_project}
\alias{devicefarm_create_project}
\title{Creates a project}
\usage{
devicefarm_create_project(name, defaultJobTimeoutMinutes)
}
\arguments{
\item{name}{[required] The project's name.}

\item{defaultJobTimeoutMinutes}{Sets the execution timeout value (in minutes) for a project. All test
runs in this project use the specified execution timeout value unless
overridden when scheduling a run.}
}
\value{
A list with the following syntax:\preformatted{list(
  project = list(
    arn = "string",
    name = "string",
    defaultJobTimeoutMinutes = 123,
    created = as.POSIXct(
      "2015-01-01"
    )
  )
)
}
}
\description{
Creates a project.
}
\section{Request syntax}{
\preformatted{svc$create_project(
  name = "string",
  defaultJobTimeoutMinutes = 123
)
}
}

\examples{
\dontrun{
# The following example creates a new project named MyProject.
svc$create_project(
  name = "MyProject"
)
}

}
\keyword{internal}
