% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_create_test_grid_project}
\alias{devicefarm_create_test_grid_project}
\title{Creates a Selenium testing project}
\usage{
devicefarm_create_test_grid_project(name, description)
}
\arguments{
\item{name}{[required] Human-readable name of the Selenium testing project.}

\item{description}{Human-readable description of the project.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  testGridProject = list(
    arn = "string",
    name = "string",
    description = "string",
    created = as.POSIXct(
      "2015-01-01"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Creates a Selenium testing project. Projects are used to track
TestGridSession instances.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_test_grid_project(
  name = "string",
  description = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
