% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_create_test_grid_url}
\alias{devicefarm_create_test_grid_url}
\title{Creates a signed, short-term URL that can be passed to a Selenium
RemoteWebDriver constructor}
\usage{
devicefarm_create_test_grid_url(projectArn, expiresInSeconds)
}
\arguments{
\item{projectArn}{[required] ARN (from
\code{\link[=devicefarm_create_test_grid_project]{create_test_grid_project}} or
\code{\link[=devicefarm_list_test_grid_projects]{list_test_grid_projects}}) to
associate with the short-term URL.}

\item{expiresInSeconds}{[required] Lifetime, in seconds, of the URL.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  url = "string",
  expires = as.POSIXct(
    "2015-01-01"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Creates a signed, short-term URL that can be passed to a Selenium
\code{RemoteWebDriver} constructor.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_test_grid_url(
  projectArn = "string",
  expiresInSeconds = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
