% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_delete_device_pool}
\alias{devicefarm_delete_device_pool}
\title{Deletes a device pool given the pool ARN}
\usage{
devicefarm_delete_device_pool(arn)
}
\arguments{
\item{arn}{[required] Represents the Amazon Resource Name (ARN) of the Device Farm device pool
to delete.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Deletes a device pool given the pool ARN. Does not allow deletion of
curated pools owned by the system.
}
\section{Request syntax}{
\preformatted{svc$delete_device_pool(
  arn = "string"
)
}
}

\examples{
\dontrun{
# The following example deletes a specific device pool.
svc$delete_device_pool(
  arn = "arn:aws:devicefarm:us-west-2::devicepool:123-456-EXAMPLE-GUID"
)
}

}
\keyword{internal}
