% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_delete_instance_profile}
\alias{devicefarm_delete_instance_profile}
\title{Deletes a profile that can be applied to one or more private device
instances}
\usage{
devicefarm_delete_instance_profile(arn)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of the instance profile you are
requesting to delete.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Deletes a profile that can be applied to one or more private device
instances.
}
\section{Request syntax}{
\preformatted{svc$delete_instance_profile(
  arn = "string"
)
}
}

\keyword{internal}
