% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_delete_network_profile}
\alias{devicefarm_delete_network_profile}
\title{Deletes a network profile}
\usage{
devicefarm_delete_network_profile(arn)
}
\arguments{
\item{arn}{[required] The ARN of the network profile to delete.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Deletes a network profile.
}
\section{Request syntax}{
\preformatted{svc$delete_network_profile(
  arn = "string"
)
}
}

\keyword{internal}
