% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_delete_project}
\alias{devicefarm_delete_project}
\title{Deletes an AWS Device Farm project, given the project ARN}
\usage{
devicefarm_delete_project(arn)
}
\arguments{
\item{arn}{[required] Represents the Amazon Resource Name (ARN) of the Device Farm project to
delete.}
}
\description{
Deletes an AWS Device Farm project, given the project ARN.

Deleting this resource does not stop an in-progress run.
}
\section{Request syntax}{
\preformatted{svc$delete_project(
  arn = "string"
)
}
}

\examples{
\dontrun{
# The following example deletes a specific project.
svc$delete_project(
  arn = "arn:aws:devicefarm:us-west-2:123456789101:project:EXAMPLE-GUID-123-456"
)
}

}
\keyword{internal}
