% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_delete_remote_access_session}
\alias{devicefarm_delete_remote_access_session}
\title{Deletes a completed remote access session and its results}
\usage{
devicefarm_delete_remote_access_session(arn)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of the session for which you want to
delete remote access.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Deletes a completed remote access session and its results.
}
\section{Request syntax}{
\preformatted{svc$delete_remote_access_session(
  arn = "string"
)
}
}

\examples{
\dontrun{
# The following example deletes a specific remote access session.
svc$delete_remote_access_session(
  arn = "arn:aws:devicefarm:us-west-2:123456789101:session:EXAMPLE-GUID-123-456"
)
}

}
\keyword{internal}
