% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_delete_run}
\alias{devicefarm_delete_run}
\title{Deletes the run, given the run ARN}
\usage{
devicefarm_delete_run(arn)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) for the run to delete.}
}
\description{
Deletes the run, given the run ARN.
}
\details{
Deleting this resource does not stop an in-progress run.
}
\section{Request syntax}{
\preformatted{svc$delete_run(
  arn = "string"
)
}
}

\section{Examples}{
\preformatted{# The following example deletes a specific test run.
svc$delete_run(
  arn = "arn:aws:devicefarm:us-west-2:123456789101:run:EXAMPLE-GUID-123-456"
)
}
}

\keyword{internal}
