% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_delete_test_grid_project}
\alias{devicefarm_delete_test_grid_project}
\title{Deletes a Selenium testing project and all content generated under it}
\usage{
devicefarm_delete_test_grid_project(projectArn)
}
\arguments{
\item{projectArn}{[required] The ARN of the project to delete, from
\code{\link[=devicefarm_create_test_grid_project]{create_test_grid_project}} or
\code{\link[=devicefarm_list_test_grid_projects]{list_test_grid_projects}}.}
}
\value{
An empty list.
}
\description{
Deletes a Selenium testing project and all content generated under it.

You cannot undo this operation.

You cannot delete a project if it has active sessions.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_test_grid_project(
  projectArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
