% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_delete_upload}
\alias{devicefarm_delete_upload}
\title{Deletes an upload given the upload ARN}
\usage{
devicefarm_delete_upload(arn)
}
\arguments{
\item{arn}{[required] Represents the Amazon Resource Name (ARN) of the Device Farm upload to
delete.}
}
\value{
An empty list.
}
\description{
Deletes an upload given the upload ARN.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_upload(
  arn = "string"
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# The following example deletes a specific upload.
svc$delete_upload(
  arn = "arn:aws:devicefarm:us-west-2:123456789101:upload:EXAMPLE-GUID-123-456"
)
}

}
\keyword{internal}
