% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_delete_vpce_configuration}
\alias{devicefarm_delete_vpce_configuration}
\title{Deletes a configuration for your Amazon Virtual Private Cloud (VPC)
endpoint}
\usage{
devicefarm_delete_vpce_configuration(arn)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of the VPC endpoint configuration you
want to delete.}
}
\description{
Deletes a configuration for your Amazon Virtual Private Cloud (VPC)
endpoint.
}
\section{Request syntax}{
\preformatted{svc$delete_vpce_configuration(
  arn = "string"
)
}
}

\keyword{internal}
