% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_get_account_settings}
\alias{devicefarm_get_account_settings}
\title{Returns the number of unmetered iOS or unmetered Android devices that
have been purchased by the account}
\usage{
devicefarm_get_account_settings()
}
\value{
A list with the following syntax:\preformatted{list(
  accountSettings = list(
    awsAccountNumber = "string",
    unmeteredDevices = list(
      123
    ),
    unmeteredRemoteAccessDevices = list(
      123
    ),
    maxJobTimeoutMinutes = 123,
    trialMinutes = list(
      total = 123.0,
      remaining = 123.0
    ),
    maxSlots = list(
      123
    ),
    defaultJobTimeoutMinutes = 123,
    skipAppResign = TRUE|FALSE
  )
)
}
}
\description{
Returns the number of unmetered iOS or unmetered Android devices that
have been purchased by the account.
}
\section{Request syntax}{
\preformatted{svc$get_account_settings()
}
}

\examples{
\dontrun{
# The following example returns information about your Device Farm account
# settings.
svc$get_account_settings()
}

}
\keyword{internal}
