% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_get_device_instance}
\alias{devicefarm_get_device_instance}
\title{Returns information about a device instance that belongs to a private
device fleet}
\usage{
devicefarm_get_device_instance(arn)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of the instance you're requesting
information about.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  deviceInstance = list(
    arn = "string",
    deviceArn = "string",
    labels = list(
      "string"
    ),
    status = "IN_USE"|"PREPARING"|"AVAILABLE"|"NOT_AVAILABLE",
    udid = "string",
    instanceProfile = list(
      arn = "string",
      packageCleanup = TRUE|FALSE,
      excludeAppPackagesFromCleanup = list(
        "string"
      ),
      rebootAfterUse = TRUE|FALSE,
      name = "string",
      description = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns information about a device instance that belongs to a private
device fleet.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_device_instance(
  arn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
