% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_get_device_pool}
\alias{devicefarm_get_device_pool}
\title{Gets information about a device pool}
\usage{
devicefarm_get_device_pool(arn)
}
\arguments{
\item{arn}{[required] The device pool's ARN.}
}
\value{
A list with the following syntax:\preformatted{list(
  devicePool = list(
    arn = "string",
    name = "string",
    description = "string",
    type = "CURATED"|"PRIVATE",
    rules = list(
      list(
        attribute = "ARN"|"PLATFORM"|"FORM_FACTOR"|"MANUFACTURER"|"REMOTE_ACCESS_ENABLED"|"REMOTE_DEBUG_ENABLED"|"APPIUM_VERSION"|"INSTANCE_ARN"|"INSTANCE_LABELS"|"FLEET_TYPE"|"OS_VERSION"|"MODEL"|"AVAILABILITY",
        operator = "EQUALS"|"LESS_THAN"|"LESS_THAN_OR_EQUALS"|"GREATER_THAN"|"GREATER_THAN_OR_EQUALS"|"IN"|"NOT_IN"|"CONTAINS",
        value = "string"
      )
    ),
    maxDevices = 123
  )
)
}
}
\description{
Gets information about a device pool.
}
\section{Request syntax}{
\preformatted{svc$get_device_pool(
  arn = "string"
)
}
}

\examples{
\dontrun{
# The following example returns information about a specific device pool,
# given a project ARN.
svc$get_device_pool(
  arn = "arn:aws:devicefarm:us-west-2:123456789101:project:EXAMPLE-GUID-123-456"
)
}

}
\keyword{internal}
