% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_get_network_profile}
\alias{devicefarm_get_network_profile}
\title{Returns information about a network profile}
\usage{
devicefarm_get_network_profile(arn)
}
\arguments{
\item{arn}{[required] The ARN of the network profile to return information about.}
}
\value{
A list with the following syntax:\preformatted{list(
  networkProfile = list(
    arn = "string",
    name = "string",
    description = "string",
    type = "CURATED"|"PRIVATE",
    uplinkBandwidthBits = 123,
    downlinkBandwidthBits = 123,
    uplinkDelayMs = 123,
    downlinkDelayMs = 123,
    uplinkJitterMs = 123,
    downlinkJitterMs = 123,
    uplinkLossPercent = 123,
    downlinkLossPercent = 123
  )
)
}
}
\description{
Returns information about a network profile.
}
\section{Request syntax}{
\preformatted{svc$get_network_profile(
  arn = "string"
)
}
}

\keyword{internal}
