% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_get_project}
\alias{devicefarm_get_project}
\title{Gets information about a project}
\usage{
devicefarm_get_project(arn)
}
\arguments{
\item{arn}{[required] The project's ARN.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  project = list(
    arn = "string",
    name = "string",
    defaultJobTimeoutMinutes = 123,
    created = as.POSIXct(
      "2015-01-01"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Gets information about a project.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_project(
  arn = "string"
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# The following example gets information about a specific project.
svc$get_project(
  arn = "arn:aws:devicefarm:us-west-2:123456789101:project:5e01a8c7-c861-4c..."
)
}

}
\keyword{internal}
