% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_get_remote_access_session}
\alias{devicefarm_get_remote_access_session}
\title{Returns a link to a currently running remote access session}
\usage{
devicefarm_get_remote_access_session(arn)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of the remote access session about which
you want to get session information.}
}
\value{
A list with the following syntax:\preformatted{list(
  remoteAccessSession = list(
    arn = "string",
    name = "string",
    created = as.POSIXct(
      "2015-01-01"
    ),
    status = "PENDING"|"PENDING_CONCURRENCY"|"PENDING_DEVICE"|"PROCESSING"|"SCHEDULING"|"PREPARING"|"RUNNING"|"COMPLETED"|"STOPPING",
    result = "PENDING"|"PASSED"|"WARNED"|"FAILED"|"SKIPPED"|"ERRORED"|"STOPPED",
    message = "string",
    started = as.POSIXct(
      "2015-01-01"
    ),
    stopped = as.POSIXct(
      "2015-01-01"
    ),
    device = list(
      arn = "string",
      name = "string",
      manufacturer = "string",
      model = "string",
      modelId = "string",
      formFactor = "PHONE"|"TABLET",
      platform = "ANDROID"|"IOS",
      os = "string",
      cpu = list(
        frequency = "string",
        architecture = "string",
        clock = 123.0
      ),
      resolution = list(
        width = 123,
        height = 123
      ),
      heapSize = 123,
      memory = 123,
      image = "string",
      carrier = "string",
      radio = "string",
      remoteAccessEnabled = TRUE|FALSE,
      remoteDebugEnabled = TRUE|FALSE,
      fleetType = "string",
      fleetName = "string",
      instances = list(
        list(
          arn = "string",
          deviceArn = "string",
          labels = list(
            "string"
          ),
          status = "IN_USE"|"PREPARING"|"AVAILABLE"|"NOT_AVAILABLE",
          udid = "string",
          instanceProfile = list(
            arn = "string",
            packageCleanup = TRUE|FALSE,
            excludeAppPackagesFromCleanup = list(
              "string"
            ),
            rebootAfterUse = TRUE|FALSE,
            name = "string",
            description = "string"
          )
        )
      ),
      availability = "TEMPORARY_NOT_AVAILABLE"|"BUSY"|"AVAILABLE"|"HIGHLY_AVAILABLE"
    ),
    instanceArn = "string",
    remoteDebugEnabled = TRUE|FALSE,
    remoteRecordEnabled = TRUE|FALSE,
    remoteRecordAppArn = "string",
    hostAddress = "string",
    clientId = "string",
    billingMethod = "METERED"|"UNMETERED",
    deviceMinutes = list(
      total = 123.0,
      metered = 123.0,
      unmetered = 123.0
    ),
    endpoint = "string",
    deviceUdid = "string",
    interactionMode = "INTERACTIVE"|"NO_VIDEO"|"VIDEO_ONLY",
    skipAppResign = TRUE|FALSE
  )
)
}
}
\description{
Returns a link to a currently running remote access session.
}
\section{Request syntax}{
\preformatted{svc$get_remote_access_session(
  arn = "string"
)
}
}

\examples{
\dontrun{
# The following example gets a specific remote access session.
svc$get_remote_access_session(
  arn = "arn:aws:devicefarm:us-west-2:123456789101:session:EXAMPLE-GUID-123-456"
)
}

}
\keyword{internal}
