% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_get_test_grid_project}
\alias{devicefarm_get_test_grid_project}
\title{Retrieves information about a Selenium testing project}
\usage{
devicefarm_get_test_grid_project(projectArn)
}
\arguments{
\item{projectArn}{[required] The ARN of the Selenium testing project, from either
\code{\link[=devicefarm_create_test_grid_project]{create_test_grid_project}} or
\code{\link[=devicefarm_list_test_grid_projects]{list_test_grid_projects}}.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  testGridProject = list(
    arn = "string",
    name = "string",
    description = "string",
    created = as.POSIXct(
      "2015-01-01"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves information about a Selenium testing project.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_test_grid_project(
  projectArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
