% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_get_test_grid_session}
\alias{devicefarm_get_test_grid_session}
\title{A session is an instance of a browser created through a RemoteWebDriver
with the URL from CreateTestGridUrlResult$url}
\usage{
devicefarm_get_test_grid_session(projectArn, sessionId, sessionArn)
}
\arguments{
\item{projectArn}{The ARN for the project that this session belongs to. See
\code{\link[=devicefarm_create_test_grid_project]{create_test_grid_project}} and
\code{\link[=devicefarm_list_test_grid_projects]{list_test_grid_projects}}.}

\item{sessionId}{An ID associated with this session.}

\item{sessionArn}{An ARN that uniquely identifies a TestGridSession.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  testGridSession = list(
    arn = "string",
    status = "ACTIVE"|"CLOSED"|"ERRORED",
    created = as.POSIXct(
      "2015-01-01"
    ),
    ended = as.POSIXct(
      "2015-01-01"
    ),
    billingMinutes = 123.0,
    seleniumProperties = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
A session is an instance of a browser created through a
\code{RemoteWebDriver} with the URL from CreateTestGridUrlResult$url. You can
use the following to look up sessions:
\itemize{
\item The session ARN (GetTestGridSessionRequest$sessionArn).
\item The project ARN and a session ID
(GetTestGridSessionRequest$projectArn and
GetTestGridSessionRequest$sessionId).
}
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_test_grid_session(
  projectArn = "string",
  sessionId = "string",
  sessionArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
