% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_list_artifacts}
\alias{devicefarm_list_artifacts}
\title{Gets information about artifacts}
\usage{
devicefarm_list_artifacts(arn, type, nextToken)
}
\arguments{
\item{arn}{[required] The run, job, suite, or test ARN.}

\item{type}{[required] The artifacts' type.

Allowed values include:
\itemize{
\item FILE
\item LOG
\item SCREENSHOT
}}

\item{nextToken}{An identifier that was returned from the previous call to this
operation, which can be used to return the next set of items in the
list.}
}
\value{
A list with the following syntax:\preformatted{list(
  artifacts = list(
    list(
      arn = "string",
      name = "string",
      type = "UNKNOWN"|"SCREENSHOT"|"DEVICE_LOG"|"MESSAGE_LOG"|"VIDEO_LOG"|"RESULT_LOG"|"SERVICE_LOG"|"WEBKIT_LOG"|"INSTRUMENTATION_OUTPUT"|"EXERCISER_MONKEY_OUTPUT"|"CALABASH_JSON_OUTPUT"|"CALABASH_PRETTY_OUTPUT"|"CALABASH_STANDARD_OUTPUT"|"CALABASH_JAVA_XML_OUTPUT"|"AUTOMATION_OUTPUT"|"APPIUM_SERVER_OUTPUT"|"APPIUM_JAVA_OUTPUT"|"APPIUM_JAVA_XML_OUTPUT"|"APPIUM_PYTHON_OUTPUT"|"APPIUM_PYTHON_XML_OUTPUT"|"EXPLORER_EVENT_LOG"|"EXPLORER_SUMMARY_LOG"|"APPLICATION_CRASH_REPORT"|"XCTEST_LOG"|"VIDEO"|"CUSTOMER_ARTIFACT"|"CUSTOMER_ARTIFACT_LOG"|"TESTSPEC_OUTPUT",
      extension = "string",
      url = "string"
    )
  ),
  nextToken = "string"
)
}
}
\description{
Gets information about artifacts.
}
\section{Request syntax}{
\preformatted{svc$list_artifacts(
  arn = "string",
  type = "SCREENSHOT"|"FILE"|"LOG",
  nextToken = "string"
)
}
}

\examples{
\dontrun{
# The following example lists screenshot artifacts for a specific run.
svc$list_artifacts(
  type = "SCREENSHOT",
  arn = "arn:aws:devicefarm:us-west-2:123456789101:run:EXAMPLE-GUID-123-456"
)
}

}
\keyword{internal}
