% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_list_device_instances}
\alias{devicefarm_list_device_instances}
\title{Returns information about the private device instances associated with
one or more AWS accounts}
\usage{
devicefarm_list_device_instances(maxResults, nextToken)
}
\arguments{
\item{maxResults}{An integer that specifies the maximum number of items you want to return
in the API response.}

\item{nextToken}{An identifier that was returned from the previous call to this
operation, which can be used to return the next set of items in the
list.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  deviceInstances = list(
    list(
      arn = "string",
      deviceArn = "string",
      labels = list(
        "string"
      ),
      status = "IN_USE"|"PREPARING"|"AVAILABLE"|"NOT_AVAILABLE",
      udid = "string",
      instanceProfile = list(
        arn = "string",
        packageCleanup = TRUE|FALSE,
        excludeAppPackagesFromCleanup = list(
          "string"
        ),
        rebootAfterUse = TRUE|FALSE,
        name = "string",
        description = "string"
      )
    )
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Returns information about the private device instances associated with
one or more AWS accounts.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_device_instances(
  maxResults = 123,
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
