% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_list_device_pools}
\alias{devicefarm_list_device_pools}
\title{Gets information about device pools}
\usage{
devicefarm_list_device_pools(arn, type, nextToken)
}
\arguments{
\item{arn}{[required] The project ARN.}

\item{type}{The device pools' type.

Allowed values include:
\itemize{
\item CURATED: A device pool that is created and managed by AWS Device
Farm.
\item PRIVATE: A device pool that is created and managed by the device
pool developer.
}}

\item{nextToken}{An identifier that was returned from the previous call to this
operation, which can be used to return the next set of items in the
list.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  devicePools = list(
    list(
      arn = "string",
      name = "string",
      description = "string",
      type = "CURATED"|"PRIVATE",
      rules = list(
        list(
          attribute = "ARN"|"PLATFORM"|"FORM_FACTOR"|"MANUFACTURER"|"REMOTE_ACCESS_ENABLED"|"REMOTE_DEBUG_ENABLED"|"APPIUM_VERSION"|"INSTANCE_ARN"|"INSTANCE_LABELS"|"FLEET_TYPE"|"OS_VERSION"|"MODEL"|"AVAILABILITY",
          operator = "EQUALS"|"LESS_THAN"|"LESS_THAN_OR_EQUALS"|"GREATER_THAN"|"GREATER_THAN_OR_EQUALS"|"IN"|"NOT_IN"|"CONTAINS",
          value = "string"
        )
      ),
      maxDevices = 123
    )
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Gets information about device pools.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_device_pools(
  arn = "string",
  type = "CURATED"|"PRIVATE",
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# The following example returns information about the private device pools
# in a specific project.
svc$list_device_pools(
  type = "PRIVATE",
  arn = "arn:aws:devicefarm:us-west-2:123456789101:project:EXAMPLE-GUID-123-456"
)
}

}
\keyword{internal}
