% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_list_network_profiles}
\alias{devicefarm_list_network_profiles}
\title{Returns the list of available network profiles}
\usage{
devicefarm_list_network_profiles(arn, type, nextToken)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of the project for which you want to list
network profiles.}

\item{type}{The type of network profile to return information about. Valid values
are listed here.}

\item{nextToken}{An identifier that was returned from the previous call to this
operation, which can be used to return the next set of items in the
list.}
}
\description{
Returns the list of available network profiles.
}
\section{Request syntax}{
\preformatted{svc$list_network_profiles(
  arn = "string",
  type = "CURATED"|"PRIVATE",
  nextToken = "string"
)
}
}

\keyword{internal}
