% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_list_offering_promotions}
\alias{devicefarm_list_offering_promotions}
\title{Returns a list of offering promotions}
\usage{
devicefarm_list_offering_promotions(nextToken)
}
\arguments{
\item{nextToken}{An identifier that was returned from the previous call to this
operation, which can be used to return the next set of items in the
list.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  offeringPromotions = list(
    list(
      id = "string",
      description = "string"
    )
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Returns a list of offering promotions. Each offering promotion record
contains the ID and description of the promotion. The API returns a
\code{NotEligible} error if the caller is not permitted to invoke the
operation. Contact \href{mailto:aws-devicefarm-support@amazon.com}{aws-devicefarm-support@amazon.com} if you must be
able to invoke this operation.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_offering_promotions(
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
