% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_list_offering_transactions}
\alias{devicefarm_list_offering_transactions}
\title{Returns a list of all historical purchases, renewals, and system renewal
transactions for an AWS account}
\usage{
devicefarm_list_offering_transactions(nextToken)
}
\arguments{
\item{nextToken}{An identifier that was returned from the previous call to this
operation, which can be used to return the next set of items in the
list.}
}
\value{
A list with the following syntax:\preformatted{list(
  offeringTransactions = list(
    list(
      offeringStatus = list(
        type = "PURCHASE"|"RENEW"|"SYSTEM",
        offering = list(
          id = "string",
          description = "string",
          type = "RECURRING",
          platform = "ANDROID"|"IOS",
          recurringCharges = list(
            list(
              cost = list(
                amount = 123.0,
                currencyCode = "USD"
              ),
              frequency = "MONTHLY"
            )
          )
        ),
        quantity = 123,
        effectiveOn = as.POSIXct(
          "2015-01-01"
        )
      ),
      transactionId = "string",
      offeringPromotionId = "string",
      createdOn = as.POSIXct(
        "2015-01-01"
      ),
      cost = list(
        amount = 123.0,
        currencyCode = "USD"
      )
    )
  ),
  nextToken = "string"
)
}
}
\description{
Returns a list of all historical purchases, renewals, and system renewal
transactions for an AWS account. The list is paginated and ordered by a
descending timestamp (most recent transactions are first). The API
returns a \code{NotEligible} error if the user is not permitted to invoke the
operation. If you must be able to invoke this operation, contact
\email{aws-devicefarm-support@amazon.com}.
}
\section{Request syntax}{
\preformatted{svc$list_offering_transactions(
  nextToken = "string"
)
}
}

\examples{
\dontrun{
# The following example returns information about Device Farm offering
# transactions.
svc$list_offering_transactions(
  nextToken = "RW5DdDJkMWYwZjM2MzM2VHVpOHJIUXlDUXlhc2QzRGViYnc9SEXAMPLE="
)
}

}
\keyword{internal}
