% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_list_offerings}
\alias{devicefarm_list_offerings}
\title{Returns a list of products or offerings that the user can manage through
the API}
\usage{
devicefarm_list_offerings(nextToken)
}
\arguments{
\item{nextToken}{An identifier that was returned from the previous call to this
operation, which can be used to return the next set of items in the
list.}
}
\value{
A list with the following syntax:\preformatted{list(
  offerings = list(
    list(
      id = "string",
      description = "string",
      type = "RECURRING",
      platform = "ANDROID"|"IOS",
      recurringCharges = list(
        list(
          cost = list(
            amount = 123.0,
            currencyCode = "USD"
          ),
          frequency = "MONTHLY"
        )
      )
    )
  ),
  nextToken = "string"
)
}
}
\description{
Returns a list of products or offerings that the user can manage through
the API. Each offering record indicates the recurring price per unit and
the frequency for that offering. The API returns a \code{NotEligible} error
if the user is not permitted to invoke the operation. If you must be
able to invoke this operation, contact
\email{aws-devicefarm-support@amazon.com}.
}
\section{Request syntax}{
\preformatted{svc$list_offerings(
  nextToken = "string"
)
}
}

\examples{
\dontrun{
# The following example returns information about available device
# offerings.
svc$list_offerings(
  nextToken = "RW5DdDJkMWYwZjM2MzM2VHVpOHJIUXlDUXlhc2QzRGViYnc9SEXAMPLE="
)
}

}
\keyword{internal}
