% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_list_projects}
\alias{devicefarm_list_projects}
\title{Gets information about projects}
\usage{
devicefarm_list_projects(arn, nextToken)
}
\arguments{
\item{arn}{Optional. If no Amazon Resource Name (ARN) is specified, then AWS Device
Farm returns a list of all projects for the AWS account. You can also
specify a project ARN.}

\item{nextToken}{An identifier that was returned from the previous call to this
operation, which can be used to return the next set of items in the
list.}
}
\value{
A list with the following syntax:\preformatted{list(
  projects = list(
    list(
      arn = "string",
      name = "string",
      defaultJobTimeoutMinutes = 123,
      created = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  nextToken = "string"
)
}
}
\description{
Gets information about projects.
}
\section{Request syntax}{
\preformatted{svc$list_projects(
  arn = "string",
  nextToken = "string"
)
}
}

\examples{
\dontrun{
# The following example returns information about the specified project in
# Device Farm.
svc$list_projects(
  arn = "arn:aws:devicefarm:us-west-2:123456789101:project:7ad300ed-8183-41...",
  nextToken = "RW5DdDJkMWYwZjM2MzM2VHVpOHJIUXlDUXlhc2QzRGViYnc9SEXAMPLE"
)
}

}
\keyword{internal}
