% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_list_remote_access_sessions}
\alias{devicefarm_list_remote_access_sessions}
\title{Returns a list of all currently running remote access sessions}
\usage{
devicefarm_list_remote_access_sessions(arn, nextToken)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of the project about which you are
requesting information.}

\item{nextToken}{An identifier that was returned from the previous call to this
operation, which can be used to return the next set of items in the
list.}
}
\value{
A list with the following syntax:\preformatted{list(
  remoteAccessSessions = list(
    list(
      arn = "string",
      name = "string",
      created = as.POSIXct(
        "2015-01-01"
      ),
      status = "PENDING"|"PENDING_CONCURRENCY"|"PENDING_DEVICE"|"PROCESSING"|"SCHEDULING"|"PREPARING"|"RUNNING"|"COMPLETED"|"STOPPING",
      result = "PENDING"|"PASSED"|"WARNED"|"FAILED"|"SKIPPED"|"ERRORED"|"STOPPED",
      message = "string",
      started = as.POSIXct(
        "2015-01-01"
      ),
      stopped = as.POSIXct(
        "2015-01-01"
      ),
      device = list(
        arn = "string",
        name = "string",
        manufacturer = "string",
        model = "string",
        modelId = "string",
        formFactor = "PHONE"|"TABLET",
        platform = "ANDROID"|"IOS",
        os = "string",
        cpu = list(
          frequency = "string",
          architecture = "string",
          clock = 123.0
        ),
        resolution = list(
          width = 123,
          height = 123
        ),
        heapSize = 123,
        memory = 123,
        image = "string",
        carrier = "string",
        radio = "string",
        remoteAccessEnabled = TRUE|FALSE,
        remoteDebugEnabled = TRUE|FALSE,
        fleetType = "string",
        fleetName = "string",
        instances = list(
          list(
            arn = "string",
            deviceArn = "string",
            labels = list(
              "string"
            ),
            status = "IN_USE"|"PREPARING"|"AVAILABLE"|"NOT_AVAILABLE",
            udid = "string",
            instanceProfile = list(
              arn = "string",
              packageCleanup = TRUE|FALSE,
              excludeAppPackagesFromCleanup = list(
                "string"
              ),
              rebootAfterUse = TRUE|FALSE,
              name = "string",
              description = "string"
            )
          )
        ),
        availability = "TEMPORARY_NOT_AVAILABLE"|"BUSY"|"AVAILABLE"|"HIGHLY_AVAILABLE"
      ),
      instanceArn = "string",
      remoteDebugEnabled = TRUE|FALSE,
      remoteRecordEnabled = TRUE|FALSE,
      remoteRecordAppArn = "string",
      hostAddress = "string",
      clientId = "string",
      billingMethod = "METERED"|"UNMETERED",
      deviceMinutes = list(
        total = 123.0,
        metered = 123.0,
        unmetered = 123.0
      ),
      endpoint = "string",
      deviceUdid = "string",
      interactionMode = "INTERACTIVE"|"NO_VIDEO"|"VIDEO_ONLY",
      skipAppResign = TRUE|FALSE
    )
  ),
  nextToken = "string"
)
}
}
\description{
Returns a list of all currently running remote access sessions.
}
\section{Request syntax}{
\preformatted{svc$list_remote_access_sessions(
  arn = "string",
  nextToken = "string"
)
}
}

\examples{
\dontrun{
# The following example returns information about a specific Device Farm
# remote access session.
svc$list_remote_access_sessions(
  arn = "arn:aws:devicefarm:us-west-2:123456789101:session:EXAMPLE-GUID-123-456",
  nextToken = "RW5DdDJkMWYwZjM2MzM2VHVpOHJIUXlDUXlhc2QzRGViYnc9SEXAMPLE="
)
}

}
\keyword{internal}
