% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_list_suites}
\alias{devicefarm_list_suites}
\title{Gets information about test suites for a given job}
\usage{
devicefarm_list_suites(arn, nextToken)
}
\arguments{
\item{arn}{[required] The job's Amazon Resource Name (ARN).}

\item{nextToken}{An identifier that was returned from the previous call to this
operation, which can be used to return the next set of items in the
list.}
}
\value{
A list with the following syntax:\preformatted{list(
  suites = list(
    list(
      arn = "string",
      name = "string",
      type = "BUILTIN_FUZZ"|"BUILTIN_EXPLORER"|"WEB_PERFORMANCE_PROFILE"|"APPIUM_JAVA_JUNIT"|"APPIUM_JAVA_TESTNG"|"APPIUM_PYTHON"|"APPIUM_NODE"|"APPIUM_RUBY"|"APPIUM_WEB_JAVA_JUNIT"|"APPIUM_WEB_JAVA_TESTNG"|"APPIUM_WEB_PYTHON"|"APPIUM_WEB_NODE"|"APPIUM_WEB_RUBY"|"CALABASH"|"INSTRUMENTATION"|"UIAUTOMATION"|"UIAUTOMATOR"|"XCTEST"|"XCTEST_UI"|"REMOTE_ACCESS_RECORD"|"REMOTE_ACCESS_REPLAY",
      created = as.POSIXct(
        "2015-01-01"
      ),
      status = "PENDING"|"PENDING_CONCURRENCY"|"PENDING_DEVICE"|"PROCESSING"|"SCHEDULING"|"PREPARING"|"RUNNING"|"COMPLETED"|"STOPPING",
      result = "PENDING"|"PASSED"|"WARNED"|"FAILED"|"SKIPPED"|"ERRORED"|"STOPPED",
      started = as.POSIXct(
        "2015-01-01"
      ),
      stopped = as.POSIXct(
        "2015-01-01"
      ),
      counters = list(
        total = 123,
        passed = 123,
        failed = 123,
        warned = 123,
        errored = 123,
        stopped = 123,
        skipped = 123
      ),
      message = "string",
      deviceMinutes = list(
        total = 123.0,
        metered = 123.0,
        unmetered = 123.0
      )
    )
  ),
  nextToken = "string"
)
}
}
\description{
Gets information about test suites for a given job.
}
\section{Request syntax}{
\preformatted{svc$list_suites(
  arn = "string",
  nextToken = "string"
)
}
}

\examples{
\dontrun{
# The following example returns information about suites, given a specific
# Device Farm job.
svc$list_suites(
  arn = "arn:aws:devicefarm:us-west-2:123456789101:job:EXAMPLE-GUID-123-456",
  nextToken = "RW5DdDJkMWYwZjM2MzM2VHVpOHJIUXlDUXlhc2QzRGViYnc9SEXAMPLE"
)
}

}
\keyword{internal}
