% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_list_test_grid_projects}
\alias{devicefarm_list_test_grid_projects}
\title{Gets a list of all Selenium testing projects in your account}
\usage{
devicefarm_list_test_grid_projects(maxResult, nextToken)
}
\arguments{
\item{maxResult}{Return no more than this number of results.}

\item{nextToken}{From a response, used to continue a paginated listing.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  testGridProjects = list(
    list(
      arn = "string",
      name = "string",
      description = "string",
      created = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Gets a list of all Selenium testing projects in your account.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_test_grid_projects(
  maxResult = 123,
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
