% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_list_test_grid_session_actions}
\alias{devicefarm_list_test_grid_session_actions}
\title{Returns a list of the actions taken in a TestGridSession}
\usage{
devicefarm_list_test_grid_session_actions(sessionArn, maxResult,
  nextToken)
}
\arguments{
\item{sessionArn}{[required] The ARN of the session to retrieve.}

\item{maxResult}{The maximum number of sessions to return per response.}

\item{nextToken}{Pagination token.}
}
\value{
A list with the following syntax:\preformatted{list(
  actions = list(
    list(
      action = "string",
      started = as.POSIXct(
        "2015-01-01"
      ),
      duration = 123,
      statusCode = "string",
      requestMethod = "string"
    )
  ),
  nextToken = "string"
)
}
}
\description{
Returns a list of the actions taken in a TestGridSession.
}
\section{Request syntax}{
\preformatted{svc$list_test_grid_session_actions(
  sessionArn = "string",
  maxResult = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
