% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_list_test_grid_sessions}
\alias{devicefarm_list_test_grid_sessions}
\title{Retrieves a list of sessions for a TestGridProject}
\usage{
devicefarm_list_test_grid_sessions(projectArn, status,
  creationTimeAfter, creationTimeBefore, endTimeAfter, endTimeBefore,
  maxResult, nextToken)
}
\arguments{
\item{projectArn}{[required] ARN of a TestGridProject.}

\item{status}{Return only sessions in this state.}

\item{creationTimeAfter}{Return only sessions created after this time.}

\item{creationTimeBefore}{Return only sessions created before this time.}

\item{endTimeAfter}{Return only sessions that ended after this time.}

\item{endTimeBefore}{Return only sessions that ended before this time.}

\item{maxResult}{Return only this many results at a time.}

\item{nextToken}{Pagination token.}
}
\value{
A list with the following syntax:\preformatted{list(
  testGridSessions = list(
    list(
      arn = "string",
      status = "ACTIVE"|"CLOSED"|"ERRORED",
      created = as.POSIXct(
        "2015-01-01"
      ),
      ended = as.POSIXct(
        "2015-01-01"
      ),
      billingMinutes = 123.0,
      seleniumProperties = "string"
    )
  ),
  nextToken = "string"
)
}
}
\description{
Retrieves a list of sessions for a TestGridProject.
}
\section{Request syntax}{
\preformatted{svc$list_test_grid_sessions(
  projectArn = "string",
  status = "ACTIVE"|"CLOSED"|"ERRORED",
  creationTimeAfter = as.POSIXct(
    "2015-01-01"
  ),
  creationTimeBefore = as.POSIXct(
    "2015-01-01"
  ),
  endTimeAfter = as.POSIXct(
    "2015-01-01"
  ),
  endTimeBefore = as.POSIXct(
    "2015-01-01"
  ),
  maxResult = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
