% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_list_unique_problems}
\alias{devicefarm_list_unique_problems}
\title{Gets information about unique problems, such as exceptions or crashes}
\usage{
devicefarm_list_unique_problems(arn, nextToken)
}
\arguments{
\item{arn}{[required] The unique problems' ARNs.}

\item{nextToken}{An identifier that was returned from the previous call to this
operation, which can be used to return the next set of items in the
list.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  uniqueProblems = list(
    list(
      list(
        message = "string",
        problems = list(
          list(
            run = list(
              arn = "string",
              name = "string"
            ),
            job = list(
              arn = "string",
              name = "string"
            ),
            suite = list(
              arn = "string",
              name = "string"
            ),
            test = list(
              arn = "string",
              name = "string"
            ),
            device = list(
              arn = "string",
              name = "string",
              manufacturer = "string",
              model = "string",
              modelId = "string",
              formFactor = "PHONE"|"TABLET",
              platform = "ANDROID"|"IOS",
              os = "string",
              cpu = list(
                frequency = "string",
                architecture = "string",
                clock = 123.0
              ),
              resolution = list(
                width = 123,
                height = 123
              ),
              heapSize = 123,
              memory = 123,
              image = "string",
              carrier = "string",
              radio = "string",
              remoteAccessEnabled = TRUE|FALSE,
              remoteDebugEnabled = TRUE|FALSE,
              fleetType = "string",
              fleetName = "string",
              instances = list(
                list(
                  arn = "string",
                  deviceArn = "string",
                  labels = list(
                    "string"
                  ),
                  status = "IN_USE"|"PREPARING"|"AVAILABLE"|"NOT_AVAILABLE",
                  udid = "string",
                  instanceProfile = list(
                    arn = "string",
                    packageCleanup = TRUE|FALSE,
                    excludeAppPackagesFromCleanup = list(
                      "string"
                    ),
                    rebootAfterUse = TRUE|FALSE,
                    name = "string",
                    description = "string"
                  )
                )
              ),
              availability = "TEMPORARY_NOT_AVAILABLE"|"BUSY"|"AVAILABLE"|"HIGHLY_AVAILABLE"
            ),
            result = "PENDING"|"PASSED"|"WARNED"|"FAILED"|"SKIPPED"|"ERRORED"|"STOPPED",
            message = "string"
          )
        )
      )
    )
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Gets information about unique problems, such as exceptions or crashes.

Unique problems are defined as a single instance of an error across a
run, job, or suite. For example, if a call in your application
consistently raises an exception
(\verb{OutOfBoundsException in MyActivity.java:386}),
\code{\link[=devicefarm_list_unique_problems]{list_unique_problems}} returns a
single entry instead of many individual entries for that exception.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_unique_problems(
  arn = "string",
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# The following example returns information about unique problems, given a
# specific Device Farm project.
svc$list_unique_problems(
  arn = "arn:aws:devicefarm:us-west-2:123456789101:project:EXAMPLE-GUID-123-456",
  nextToken = "RW5DdDJkMWYwZjM2MzM2VHVpOHJIUXlDUXlhc2QzRGViYnc9SEXAMPLE"
)
}

}
\keyword{internal}
