% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_purchase_offering}
\alias{devicefarm_purchase_offering}
\title{Immediately purchases offerings for an AWS account}
\usage{
devicefarm_purchase_offering(offeringId, quantity, offeringPromotionId)
}
\arguments{
\item{offeringId}{The ID of the offering.}

\item{quantity}{The number of device slots to purchase in an offering request.}

\item{offeringPromotionId}{The ID of the offering promotion to be applied to the purchase.}
}
\value{
A list with the following syntax:\preformatted{list(
  offeringTransaction = list(
    offeringStatus = list(
      type = "PURCHASE"|"RENEW"|"SYSTEM",
      offering = list(
        id = "string",
        description = "string",
        type = "RECURRING",
        platform = "ANDROID"|"IOS",
        recurringCharges = list(
          list(
            cost = list(
              amount = 123.0,
              currencyCode = "USD"
            ),
            frequency = "MONTHLY"
          )
        )
      ),
      quantity = 123,
      effectiveOn = as.POSIXct(
        "2015-01-01"
      )
    ),
    transactionId = "string",
    offeringPromotionId = "string",
    createdOn = as.POSIXct(
      "2015-01-01"
    ),
    cost = list(
      amount = 123.0,
      currencyCode = "USD"
    )
  )
)
}
}
\description{
Immediately purchases offerings for an AWS account. Offerings renew with
the latest total purchased quantity for an offering, unless the renewal
was overridden. The API returns a \code{NotEligible} error if the user is not
permitted to invoke the operation. If you must be able to invoke this
operation, contact \email{aws-devicefarm-support@amazon.com}.
}
\section{Request syntax}{
\preformatted{svc$purchase_offering(
  offeringId = "string",
  quantity = 123,
  offeringPromotionId = "string"
)
}
}

\examples{
\dontrun{
# The following example purchases a specific device slot offering.
svc$purchase_offering(
  offeringId = "D68B3C05-1BA6-4360-BC69-12345EXAMPLE",
  quantity = 1L
)
}

}
\keyword{internal}
