% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_renew_offering}
\alias{devicefarm_renew_offering}
\title{Explicitly sets the quantity of devices to renew for an offering,
starting from the effectiveDate of the next period}
\usage{
devicefarm_renew_offering(offeringId, quantity)
}
\arguments{
\item{offeringId}{The ID of a request to renew an offering.}

\item{quantity}{The quantity requested in an offering renewal.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  offeringTransaction = list(
    offeringStatus = list(
      type = "PURCHASE"|"RENEW"|"SYSTEM",
      offering = list(
        id = "string",
        description = "string",
        type = "RECURRING",
        platform = "ANDROID"|"IOS",
        recurringCharges = list(
          list(
            cost = list(
              amount = 123.0,
              currencyCode = "USD"
            ),
            frequency = "MONTHLY"
          )
        )
      ),
      quantity = 123,
      effectiveOn = as.POSIXct(
        "2015-01-01"
      )
    ),
    transactionId = "string",
    offeringPromotionId = "string",
    createdOn = as.POSIXct(
      "2015-01-01"
    ),
    cost = list(
      amount = 123.0,
      currencyCode = "USD"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Explicitly sets the quantity of devices to renew for an offering,
starting from the \code{effectiveDate} of the next period. The API returns a
\code{NotEligible} error if the user is not permitted to invoke the
operation. If you must be able to invoke this operation, contact
\href{mailto:aws-devicefarm-support@amazon.com}{aws-devicefarm-support@amazon.com}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$renew_offering(
  offeringId = "string",
  quantity = 123
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# The following example renews a specific device slot offering.
svc$renew_offering(
  offeringId = "D68B3C05-1BA6-4360-BC69-12345EXAMPLE",
  quantity = 1L
)
}

}
\keyword{internal}
