% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_stop_job}
\alias{devicefarm_stop_job}
\title{Initiates a stop request for the current job}
\usage{
devicefarm_stop_job(arn)
}
\arguments{
\item{arn}{[required] Represents the Amazon Resource Name (ARN) of the Device Farm job to
stop.}
}
\description{
Initiates a stop request for the current job. AWS Device Farm
immediately stops the job on the device where tests have not started.
You are not billed for this device. On the device where tests have
started, setup suite and teardown suite tests run to completion on the
device. You are billed for setup, teardown, and any tests that were in
progress or already completed.
}
\section{Request syntax}{
\preformatted{svc$stop_job(
  arn = "string"
)
}
}

\keyword{internal}
