% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_stop_remote_access_session}
\alias{devicefarm_stop_remote_access_session}
\title{Ends a specified remote access session}
\usage{
devicefarm_stop_remote_access_session(arn)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of the remote access session you wish to
stop.}
}
\description{
Ends a specified remote access session.
}
\section{Request syntax}{
\preformatted{svc$stop_remote_access_session(
  arn = "string"
)
}
}

\keyword{internal}
