% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_update_instance_profile}
\alias{devicefarm_update_instance_profile}
\title{Updates information about an existing private device instance profile}
\usage{
devicefarm_update_instance_profile(arn, name, description,
  packageCleanup, excludeAppPackagesFromCleanup, rebootAfterUse)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of the instance profile.}

\item{name}{The updated name for your instance profile.}

\item{description}{The updated description for your instance profile.}

\item{packageCleanup}{The updated choice for whether you want to specify package cleanup. The
default value is \code{false} for private devices.}

\item{excludeAppPackagesFromCleanup}{An array of strings that specifies the list of app packages that should
not be cleaned up from the device after a test run is over.

The list of packages is only considered if you set \code{packageCleanup} to
\code{true}.}

\item{rebootAfterUse}{The updated choice for whether you want to reboot the device after use.
The default value is \code{true}.}
}
\value{
A list with the following syntax:\preformatted{list(
  instanceProfile = list(
    arn = "string",
    packageCleanup = TRUE|FALSE,
    excludeAppPackagesFromCleanup = list(
      "string"
    ),
    rebootAfterUse = TRUE|FALSE,
    name = "string",
    description = "string"
  )
)
}
}
\description{
Updates information about an existing private device instance profile.
}
\section{Request syntax}{
\preformatted{svc$update_instance_profile(
  arn = "string",
  name = "string",
  description = "string",
  packageCleanup = TRUE|FALSE,
  excludeAppPackagesFromCleanup = list(
    "string"
  ),
  rebootAfterUse = TRUE|FALSE
)
}
}

\keyword{internal}
