% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_update_project}
\alias{devicefarm_update_project}
\title{Modifies the specified project name, given the project ARN and a new
name}
\usage{
devicefarm_update_project(arn, name, defaultJobTimeoutMinutes)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of the project whose name to update.}

\item{name}{A string that represents the new name of the project that you are
updating.}

\item{defaultJobTimeoutMinutes}{The number of minutes a test run in the project executes before it times
out.}
}
\value{
A list with the following syntax:\preformatted{list(
  project = list(
    arn = "string",
    name = "string",
    defaultJobTimeoutMinutes = 123,
    created = as.POSIXct(
      "2015-01-01"
    )
  )
)
}
}
\description{
Modifies the specified project name, given the project ARN and a new
name.
}
\section{Request syntax}{
\preformatted{svc$update_project(
  arn = "string",
  name = "string",
  defaultJobTimeoutMinutes = 123
)
}
}

\examples{
\dontrun{
# The following example updates the specified project with a new name.
svc$update_project(
  name = "NewName",
  arn = "arn:aws:devicefarm:us-west-2:123456789101:project:8f75187d-101e-4625-accc-12..."
)
}

}
\keyword{internal}
