% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_update_upload}
\alias{devicefarm_update_upload}
\title{Updates an uploaded test spec}
\usage{
devicefarm_update_upload(arn, name, contentType, editContent)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of the uploaded test spec.}

\item{name}{The upload's test spec file name. The name must not contain any forward
slashes (/). The test spec file name must end with the \code{.yaml} or \code{.yml}
file extension.}

\item{contentType}{The upload's content type (for example, \code{application/x-yaml}).}

\item{editContent}{Set to true if the YAML file has changed and must be updated. Otherwise,
set to false.}
}
\value{
A list with the following syntax:\preformatted{list(
  upload = list(
    arn = "string",
    name = "string",
    created = as.POSIXct(
      "2015-01-01"
    ),
    type = "ANDROID_APP"|"IOS_APP"|"WEB_APP"|"EXTERNAL_DATA"|"APPIUM_JAVA_JUNIT_TEST_PACKAGE"|"APPIUM_JAVA_TESTNG_TEST_PACKAGE"|"APPIUM_PYTHON_TEST_PACKAGE"|"APPIUM_NODE_TEST_PACKAGE"|"APPIUM_RUBY_TEST_PACKAGE"|"APPIUM_WEB_JAVA_JUNIT_TEST_PACKAGE"|"APPIUM_WEB_JAVA_TESTNG_TEST_PACKAGE"|"APPIUM_WEB_PYTHON_TEST_PACKAGE"|"APPIUM_WEB_NODE_TEST_PACKAGE"|"APPIUM_WEB_RUBY_TEST_PACKAGE"|"CALABASH_TEST_PACKAGE"|"INSTRUMENTATION_TEST_PACKAGE"|"UIAUTOMATION_TEST_PACKAGE"|"UIAUTOMATOR_TEST_PACKAGE"|"XCTEST_TEST_PACKAGE"|"XCTEST_UI_TEST_PACKAGE"|"APPIUM_JAVA_JUNIT_TEST_SPEC"|"APPIUM_JAVA_TESTNG_TEST_SPEC"|"APPIUM_PYTHON_TEST_SPEC"|"APPIUM_NODE_TEST_SPEC"|"APPIUM_RUBY_TEST_SPEC"|"APPIUM_WEB_JAVA_JUNIT_TEST_SPEC"|"APPIUM_WEB_JAVA_TESTNG_TEST_SPEC"|"APPIUM_WEB_PYTHON_TEST_SPEC"|"APPIUM_WEB_NODE_TEST_SPEC"|"APPIUM_WEB_RUBY_TEST_SPEC"|"INSTRUMENTATION_TEST_SPEC"|"XCTEST_UI_TEST_SPEC",
    status = "INITIALIZED"|"PROCESSING"|"SUCCEEDED"|"FAILED",
    url = "string",
    metadata = "string",
    contentType = "string",
    message = "string",
    category = "CURATED"|"PRIVATE"
  )
)
}
}
\description{
Updates an uploaded test spec.
}
\section{Request syntax}{
\preformatted{svc$update_upload(
  arn = "string",
  name = "string",
  contentType = "string",
  editContent = TRUE|FALSE
)
}
}

\keyword{internal}
