% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mobile_operations.R
\name{export_bundle}
\alias{export_bundle}
\title{Generates customized software development kit (SDK) and or tool packages used to integrate mobile web or mobile app clients with backend AWS resources}
\usage{
export_bundle(bundleId, projectId = NULL, platform = NULL)
}
\arguments{
\item{bundleId}{[required] Unique bundle identifier.}

\item{projectId}{Unique project identifier.}

\item{platform}{Developer desktop or target application platform.}
}
\description{
Generates customized software development kit (SDK) and or tool packages used to integrate mobile web or mobile app clients with backend AWS resources.
}
\section{Accepted Parameters}{
\preformatted{export_bundle(
  bundleId = "string",
  projectId = "string",
  platform = "OSX"|"WINDOWS"|"LINUX"|"OBJC"|"SWIFT"|"ANDROID"|"JAVASCRIPT"
)
}
}

