% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mobile_operations.R
\name{export_project}
\alias{export_project}
\title{Exports project configuration to a snapshot which can be downloaded and shared}
\usage{
export_project(projectId)
}
\arguments{
\item{projectId}{Unique project identifier.}
}
\description{
Exports project configuration to a snapshot which can be downloaded and shared. Note that mobile app push credentials are encrypted in exported projects, so they can only be shared successfully within the same AWS account.
}
\examples{

}
