% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mobile_operations.R
\name{list_projects}
\alias{list_projects}
\title{Lists projects in AWS Mobile Hub}
\usage{
list_projects(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{Maximum number of records to list in a single response.}

\item{nextToken}{Pagination token. Set to null to start listing projects from start. If non-null pagination token is returned in a result, then pass its value in here in another request to list more projects.}
}
\description{
Lists projects in AWS Mobile Hub.
}
\examples{

}
