% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobile_operations.R
\name{mobile_create_project}
\alias{mobile_create_project}
\title{Creates an AWS Mobile Hub project}
\usage{
mobile_create_project(name, region, contents, snapshotId)
}
\arguments{
\item{name}{Name of the project.}

\item{region}{Default region where project resources should be created.}

\item{contents}{ZIP or YAML file which contains configuration settings to be used when
creating the project. This may be the contents of the file downloaded
from the URL provided in an export project operation.}

\item{snapshotId}{Unique identifier for an exported snapshot of project configuration.
This snapshot identifier is included in the share URL when a project is
exported.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  details = list(
    name = "string",
    projectId = "string",
    region = "string",
    state = "NORMAL"|"SYNCING"|"IMPORTING",
    createdDate = as.POSIXct(
      "2015-01-01"
    ),
    lastUpdatedDate = as.POSIXct(
      "2015-01-01"
    ),
    consoleUrl = "string",
    resources = list(
      list(
        type = "string",
        name = "string",
        arn = "string",
        feature = "string",
        attributes = list(
          "string"
        )
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Creates an AWS Mobile Hub project.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_project(
  name = "string",
  region = "string",
  contents = raw,
  snapshotId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
