% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobile_operations.R
\name{mobile_describe_bundle}
\alias{mobile_describe_bundle}
\title{Get the bundle details for the requested bundle id}
\usage{
mobile_describe_bundle(bundleId)
}
\arguments{
\item{bundleId}{[required] Unique bundle identifier.}
}
\value{
A list with the following syntax:\preformatted{list(
  details = list(
    bundleId = "string",
    title = "string",
    version = "string",
    description = "string",
    iconUrl = "string",
    availablePlatforms = list(
      "OSX"|"WINDOWS"|"LINUX"|"OBJC"|"SWIFT"|"ANDROID"|"JAVASCRIPT"
    )
  )
)
}
}
\description{
Get the bundle details for the requested bundle id.
}
\section{Request syntax}{
\preformatted{svc$describe_bundle(
  bundleId = "string"
)
}
}

\keyword{internal}
