% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobile_operations.R
\name{mobile_export_bundle}
\alias{mobile_export_bundle}
\title{Generates customized software development kit (SDK) and or tool packages
used to integrate mobile web or mobile app clients with backend AWS
resources}
\usage{
mobile_export_bundle(bundleId, projectId, platform)
}
\arguments{
\item{bundleId}{[required] Unique bundle identifier.}

\item{projectId}{Unique project identifier.}

\item{platform}{Developer desktop or target application platform.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  downloadUrl = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Generates customized software development kit (SDK) and or tool packages
used to integrate mobile web or mobile app clients with backend AWS
resources.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$export_bundle(
  bundleId = "string",
  projectId = "string",
  platform = "OSX"|"WINDOWS"|"LINUX"|"OBJC"|"SWIFT"|"ANDROID"|"JAVASCRIPT"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
