% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobile_operations.R
\name{mobile_export_project}
\alias{mobile_export_project}
\title{Exports project configuration to a snapshot which can be downloaded and
shared}
\usage{
mobile_export_project(projectId)
}
\arguments{
\item{projectId}{[required] Unique project identifier.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  downloadUrl = "string",
  shareUrl = "string",
  snapshotId = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Exports project configuration to a snapshot which can be downloaded and
shared. Note that mobile app push credentials are encrypted in exported
projects, so they can only be shared successfully within the same AWS
account.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$export_project(
  projectId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
