% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobile_operations.R
\name{mobile_list_bundles}
\alias{mobile_list_bundles}
\title{List all available bundles}
\usage{
mobile_list_bundles(maxResults, nextToken)
}
\arguments{
\item{maxResults}{Maximum number of records to list in a single response.}

\item{nextToken}{Pagination token. Set to null to start listing bundles from start. If
non-null pagination token is returned in a result, then pass its value
in here in another request to list more bundles.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  bundleList = list(
    list(
      bundleId = "string",
      title = "string",
      version = "string",
      description = "string",
      iconUrl = "string",
      availablePlatforms = list(
        "OSX"|"WINDOWS"|"LINUX"|"OBJC"|"SWIFT"|"ANDROID"|"JAVASCRIPT"
      )
    )
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
List all available bundles.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_bundles(
  maxResults = 123,
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
