% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobile_operations.R
\name{mobile_update_project}
\alias{mobile_update_project}
\title{Update an existing project}
\usage{
mobile_update_project(contents, projectId)
}
\arguments{
\item{contents}{ZIP or YAML file which contains project configuration to be updated.
This should be the contents of the file downloaded from the URL provided
in an export project operation.}

\item{projectId}{[required] Unique project identifier.}
}
\value{
A list with the following syntax:\preformatted{list(
  details = list(
    name = "string",
    projectId = "string",
    region = "string",
    state = "NORMAL"|"SYNCING"|"IMPORTING",
    createdDate = as.POSIXct(
      "2015-01-01"
    ),
    lastUpdatedDate = as.POSIXct(
      "2015-01-01"
    ),
    consoleUrl = "string",
    resources = list(
      list(
        type = "string",
        name = "string",
        arn = "string",
        feature = "string",
        attributes = list(
          "string"
        )
      )
    )
  )
)
}
}
\description{
Update an existing project.
}
\section{Request syntax}{
\preformatted{svc$update_project(
  contents = raw,
  projectId = "string"
)
}
}

\keyword{internal}
