% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobileanalytics_operations.R
\name{mobileanalytics_put_events}
\alias{mobileanalytics_put_events}
\title{The PutEvents operation records one or more events}
\usage{
mobileanalytics_put_events(events, clientContext, clientContextEncoding)
}
\arguments{
\item{events}{[required] An array of Event JSON objects}

\item{clientContext}{[required] The client context including the client ID, app title, app version and
package name.}

\item{clientContextEncoding}{The encoding used for the client context.}
}
\value{
An empty list.
}
\description{
The PutEvents operation records one or more events. You can have up to
1,500 unique custom events per app, any combination of up to 40
attributes and metrics per custom event, and any number of attribute or
metric values.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$put_events(
  events = list(
    list(
      eventType = "string",
      timestamp = "string",
      session = list(
        id = "string",
        duration = 123,
        startTimestamp = "string",
        stopTimestamp = "string"
      ),
      version = "string",
      attributes = list(
        "string"
      ),
      metrics = list(
        123.0
      )
    )
  ),
  clientContext = "string",
  clientContextEncoding = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
