% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mobile_package.R
\docType{package}
\name{paws.mobile-package}
\alias{paws.mobile}
\alias{paws.mobile-package}
\title{paws.mobile: AWS Mobile}
\description{
AWS Mobile Service provides mobile app and website
developers with capabilities required to configure AWS resources and
bootstrap their developer desktop projects with the necessary SDKs,
constants, tools and samples to make use of those resources.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
