% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_create_backend_environment}
\alias{amplify_create_backend_environment}
\title{Creates a new backend environment for an Amplify app}
\usage{
amplify_create_backend_environment(appId, environmentName, stackName,
  deploymentArtifacts)
}
\arguments{
\item{appId}{[required] The unique ID for an Amplify app.}

\item{environmentName}{[required] The name for the backend environment.}

\item{stackName}{The AWS CloudFormation stack name of a backend environment.}

\item{deploymentArtifacts}{The name of deployment artifacts.}
}
\value{
A list with the following syntax:\preformatted{list(
  backendEnvironment = list(
    backendEnvironmentArn = "string",
    environmentName = "string",
    stackName = "string",
    deploymentArtifacts = "string",
    createTime = as.POSIXct(
      "2015-01-01"
    ),
    updateTime = as.POSIXct(
      "2015-01-01"
    )
  )
)
}
}
\description{
Creates a new backend environment for an Amplify app.
}
\section{Request syntax}{
\preformatted{svc$create_backend_environment(
  appId = "string",
  environmentName = "string",
  stackName = "string",
  deploymentArtifacts = "string"
)
}
}

\keyword{internal}
