% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_create_deployment}
\alias{amplify_create_deployment}
\title{Creates a deployment for a manually deployed Amplify app}
\usage{
amplify_create_deployment(appId, branchName, fileMap)
}
\arguments{
\item{appId}{[required] The unique ID for an Amplify app.}

\item{branchName}{[required] The name for the branch, for the job.}

\item{fileMap}{An optional file map that contains the file name as the key and the file
content md5 hash as the value. If this argument is provided, the service
will generate a unique upload URL per file. Otherwise, the service will
only generate a single upload URL for the zipped files.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  jobId = "string",
  fileUploadUrls = list(
    "string"
  ),
  zipUploadUrl = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Creates a deployment for a manually deployed Amplify app. Manually
deployed apps are not connected to a repository.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_deployment(
  appId = "string",
  branchName = "string",
  fileMap = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
