% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_create_webhook}
\alias{amplify_create_webhook}
\title{Creates a new webhook on an Amplify app}
\usage{
amplify_create_webhook(appId, branchName, description)
}
\arguments{
\item{appId}{[required] The unique ID for an Amplify app.}

\item{branchName}{[required] The name for a branch that is part of an Amplify app.}

\item{description}{The description for a webhook.}
}
\value{
A list with the following syntax:\preformatted{list(
  webhook = list(
    webhookArn = "string",
    webhookId = "string",
    webhookUrl = "string",
    branchName = "string",
    description = "string",
    createTime = as.POSIXct(
      "2015-01-01"
    ),
    updateTime = as.POSIXct(
      "2015-01-01"
    )
  )
)
}
}
\description{
Creates a new webhook on an Amplify app.
}
\section{Request syntax}{
\preformatted{svc$create_webhook(
  appId = "string",
  branchName = "string",
  description = "string"
)
}
}

\keyword{internal}
