% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_delete_app}
\alias{amplify_delete_app}
\title{Delete an existing Amplify App by appId}
\usage{
amplify_delete_app(appId)
}
\arguments{
\item{appId}{[required] Unique Id for an Amplify App.}
}
\description{
Delete an existing Amplify App by appId.
}
\section{Request syntax}{
\preformatted{svc$delete_app(
  appId = "string"
)
}
}

\keyword{internal}
