% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_delete_backend_environment}
\alias{amplify_delete_backend_environment}
\title{Delete backend environment for an Amplify App}
\usage{
amplify_delete_backend_environment(appId, environmentName)
}
\arguments{
\item{appId}{[required] Unique Id of an Amplify App.}

\item{environmentName}{[required] Name of a backend environment of an Amplify App.}
}
\description{
Delete backend environment for an Amplify App.
}
\section{Request syntax}{
\preformatted{svc$delete_backend_environment(
  appId = "string",
  environmentName = "string"
)
}
}

\keyword{internal}
