% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_delete_webhook}
\alias{amplify_delete_webhook}
\title{Deletes a webhook}
\usage{
amplify_delete_webhook(webhookId)
}
\arguments{
\item{webhookId}{[required] The unique ID for a webhook.}
}
\description{
Deletes a webhook.
}
\section{Request syntax}{
\preformatted{svc$delete_webhook(
  webhookId = "string"
)
}
}

\keyword{internal}
