% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_get_app}
\alias{amplify_get_app}
\title{Returns an existing Amplify app by appID}
\usage{
amplify_get_app(appId)
}
\arguments{
\item{appId}{[required] The unique ID for an Amplify app.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  app = list(
    appId = "string",
    appArn = "string",
    name = "string",
    tags = list(
      "string"
    ),
    description = "string",
    repository = "string",
    platform = "WEB",
    createTime = as.POSIXct(
      "2015-01-01"
    ),
    updateTime = as.POSIXct(
      "2015-01-01"
    ),
    iamServiceRoleArn = "string",
    environmentVariables = list(
      "string"
    ),
    defaultDomain = "string",
    enableBranchAutoBuild = TRUE|FALSE,
    enableBranchAutoDeletion = TRUE|FALSE,
    enableBasicAuth = TRUE|FALSE,
    basicAuthCredentials = "string",
    customRules = list(
      list(
        source = "string",
        target = "string",
        status = "string",
        condition = "string"
      )
    ),
    productionBranch = list(
      lastDeployTime = as.POSIXct(
        "2015-01-01"
      ),
      status = "string",
      thumbnailUrl = "string",
      branchName = "string"
    ),
    buildSpec = "string",
    customHeaders = "string",
    enableAutoBranchCreation = TRUE|FALSE,
    autoBranchCreationPatterns = list(
      "string"
    ),
    autoBranchCreationConfig = list(
      stage = "PRODUCTION"|"BETA"|"DEVELOPMENT"|"EXPERIMENTAL"|"PULL_REQUEST",
      framework = "string",
      enableAutoBuild = TRUE|FALSE,
      environmentVariables = list(
        "string"
      ),
      basicAuthCredentials = "string",
      enableBasicAuth = TRUE|FALSE,
      enablePerformanceMode = TRUE|FALSE,
      buildSpec = "string",
      enablePullRequestPreview = TRUE|FALSE,
      pullRequestEnvironmentName = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns an existing Amplify app by appID.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_app(
  appId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
