% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_get_app}
\alias{amplify_get_app}
\title{Retrieves an existing Amplify App by appId}
\usage{
amplify_get_app(appId)
}
\arguments{
\item{appId}{[required] Unique Id for an Amplify App.}
}
\description{
Retrieves an existing Amplify App by appId.
}
\section{Request syntax}{
\preformatted{svc$get_app(
  appId = "string"
)
}
}

\keyword{internal}
