% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_get_artifact_url}
\alias{amplify_get_artifact_url}
\title{Retrieves artifact info that corresponds to a artifactId}
\usage{
amplify_get_artifact_url(artifactId)
}
\arguments{
\item{artifactId}{[required] Unique Id for a artifact.}
}
\description{
Retrieves artifact info that corresponds to a artifactId.
}
\section{Request syntax}{
\preformatted{svc$get_artifact_url(
  artifactId = "string"
)
}
}

\keyword{internal}
