% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_get_job}
\alias{amplify_get_job}
\title{Returns a job for a branch of an Amplify app}
\usage{
amplify_get_job(appId, branchName, jobId)
}
\arguments{
\item{appId}{[required] The unique ID for an Amplify app.}

\item{branchName}{[required] The branch name for the job.}

\item{jobId}{[required] The unique ID for the job.}
}
\value{
A list with the following syntax:\preformatted{list(
  job = list(
    summary = list(
      jobArn = "string",
      jobId = "string",
      commitId = "string",
      commitMessage = "string",
      commitTime = as.POSIXct(
        "2015-01-01"
      ),
      startTime = as.POSIXct(
        "2015-01-01"
      ),
      status = "PENDING"|"PROVISIONING"|"RUNNING"|"FAILED"|"SUCCEED"|"CANCELLING"|"CANCELLED",
      endTime = as.POSIXct(
        "2015-01-01"
      ),
      jobType = "RELEASE"|"RETRY"|"MANUAL"|"WEB_HOOK"
    ),
    steps = list(
      list(
        stepName = "string",
        startTime = as.POSIXct(
          "2015-01-01"
        ),
        status = "PENDING"|"PROVISIONING"|"RUNNING"|"FAILED"|"SUCCEED"|"CANCELLING"|"CANCELLED",
        endTime = as.POSIXct(
          "2015-01-01"
        ),
        logUrl = "string",
        artifactsUrl = "string",
        testArtifactsUrl = "string",
        testConfigUrl = "string",
        screenshots = list(
          "string"
        ),
        statusReason = "string",
        context = "string"
      )
    )
  )
)
}
}
\description{
Returns a job for a branch of an Amplify app.
}
\section{Request syntax}{
\preformatted{svc$get_job(
  appId = "string",
  branchName = "string",
  jobId = "string"
)
}
}

\keyword{internal}
