% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_list_apps}
\alias{amplify_list_apps}
\title{Lists existing Amplify Apps}
\usage{
amplify_list_apps(nextToken, maxResults)
}
\arguments{
\item{nextToken}{Pagination token. If non-null pagination token is returned in a result,
then pass its value in another request to fetch more entries.}

\item{maxResults}{Maximum number of records to list in a single response.}
}
\description{
Lists existing Amplify Apps.
}
\section{Request syntax}{
\preformatted{svc$list_apps(
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
