% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_list_artifacts}
\alias{amplify_list_artifacts}
\title{Returns a list of artifacts for a specified app, branch, and job}
\usage{
amplify_list_artifacts(appId, branchName, jobId, nextToken, maxResults)
}
\arguments{
\item{appId}{[required] The unique ID for an Amplify app.}

\item{branchName}{[required] The name of a branch that is part of an Amplify app.}

\item{jobId}{[required] The unique ID for a job.}

\item{nextToken}{A pagination token. Set to null to start listing artifacts from start.
If a non-null pagination token is returned in a result, pass its value
in here to list more artifacts.}

\item{maxResults}{The maximum number of records to list in a single response.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  artifacts = list(
    list(
      artifactFileName = "string",
      artifactId = "string"
    )
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Returns a list of artifacts for a specified app, branch, and job.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_artifacts(
  appId = "string",
  branchName = "string",
  jobId = "string",
  nextToken = "string",
  maxResults = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
