% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_list_backend_environments}
\alias{amplify_list_backend_environments}
\title{Lists the backend environments for an Amplify app}
\usage{
amplify_list_backend_environments(appId, environmentName, nextToken,
  maxResults)
}
\arguments{
\item{appId}{[required] The unique ID for an Amplify app.}

\item{environmentName}{The name of the backend environment}

\item{nextToken}{A pagination token. Set to null to start listing backend environments
from the start. If a non-null pagination token is returned in a result,
pass its value in here to list more backend environments.}

\item{maxResults}{The maximum number of records to list in a single response.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  backendEnvironments = list(
    list(
      backendEnvironmentArn = "string",
      environmentName = "string",
      stackName = "string",
      deploymentArtifacts = "string",
      createTime = as.POSIXct(
        "2015-01-01"
      ),
      updateTime = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists the backend environments for an Amplify app.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_backend_environments(
  appId = "string",
  environmentName = "string",
  nextToken = "string",
  maxResults = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
