% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_list_backend_environments}
\alias{amplify_list_backend_environments}
\title{Lists the backend environments for an Amplify app}
\usage{
amplify_list_backend_environments(appId, environmentName, nextToken,
  maxResults)
}
\arguments{
\item{appId}{[required] The unique ID for an Amplify app.}

\item{environmentName}{The name of the backend environment}

\item{nextToken}{A pagination token. Set to null to start listing backend environments
from the start. If a non-null pagination token is returned in a result,
pass its value in here to list more backend environments.}

\item{maxResults}{The maximum number of records to list in a single response.}
}
\description{
Lists the backend environments for an Amplify app.
}
\section{Request syntax}{
\preformatted{svc$list_backend_environments(
  appId = "string",
  environmentName = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
