% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_list_branches}
\alias{amplify_list_branches}
\title{Lists the branches of an Amplify app}
\usage{
amplify_list_branches(appId, nextToken, maxResults)
}
\arguments{
\item{appId}{[required] The unique ID for an Amplify app.}

\item{nextToken}{A pagination token. Set to null to start listing branches from the
start. If a non-null pagination token is returned in a result, pass its
value in here to list more branches.}

\item{maxResults}{The maximum number of records to list in a single response.}
}
\value{
A list with the following syntax:\preformatted{list(
  branches = list(
    list(
      branchArn = "string",
      branchName = "string",
      description = "string",
      tags = list(
        "string"
      ),
      stage = "PRODUCTION"|"BETA"|"DEVELOPMENT"|"EXPERIMENTAL"|"PULL_REQUEST",
      displayName = "string",
      enableNotification = TRUE|FALSE,
      createTime = as.POSIXct(
        "2015-01-01"
      ),
      updateTime = as.POSIXct(
        "2015-01-01"
      ),
      environmentVariables = list(
        "string"
      ),
      enableAutoBuild = TRUE|FALSE,
      customDomains = list(
        "string"
      ),
      framework = "string",
      activeJobId = "string",
      totalNumberOfJobs = "string",
      enableBasicAuth = TRUE|FALSE,
      enablePerformanceMode = TRUE|FALSE,
      thumbnailUrl = "string",
      basicAuthCredentials = "string",
      buildSpec = "string",
      ttl = "string",
      associatedResources = list(
        "string"
      ),
      enablePullRequestPreview = TRUE|FALSE,
      pullRequestEnvironmentName = "string",
      destinationBranch = "string",
      sourceBranch = "string",
      backendEnvironmentArn = "string"
    )
  ),
  nextToken = "string"
)
}
}
\description{
Lists the branches of an Amplify app.
}
\section{Request syntax}{
\preformatted{svc$list_branches(
  appId = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
