% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_list_jobs}
\alias{amplify_list_jobs}
\title{Lists the jobs for a branch of an Amplify app}
\usage{
amplify_list_jobs(appId, branchName, nextToken, maxResults)
}
\arguments{
\item{appId}{[required] The unique ID for an Amplify app.}

\item{branchName}{[required] The name for a branch.}

\item{nextToken}{A pagination token. Set to null to start listing steps from the start.
If a non-null pagination token is returned in a result, pass its value
in here to list more steps.}

\item{maxResults}{The maximum number of records to list in a single response.}
}
\description{
Lists the jobs for a branch of an Amplify app.
}
\section{Request syntax}{
\preformatted{svc$list_jobs(
  appId = "string",
  branchName = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
