% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_start_deployment}
\alias{amplify_start_deployment}
\title{Start a deployment for manual deploy apps}
\usage{
amplify_start_deployment(appId, branchName, jobId, sourceUrl)
}
\arguments{
\item{appId}{[required] Unique Id for an Amplify App.}

\item{branchName}{[required] Name for the branch, for the Job.}

\item{jobId}{The job id for this deployment, generated by create deployment request.}

\item{sourceUrl}{The sourceUrl for this deployment, used when calling start deployment
without create deployment. SourceUrl can be any HTTP GET url that is
public accessible and downloads a single zip.}
}
\description{
Start a deployment for manual deploy apps. (Apps are not connected to
repository)
}
\section{Request syntax}{
\preformatted{svc$start_deployment(
  appId = "string",
  branchName = "string",
  jobId = "string",
  sourceUrl = "string"
)
}
}

\keyword{internal}
