% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_start_deployment}
\alias{amplify_start_deployment}
\title{Starts a deployment for a manually deployed app}
\usage{
amplify_start_deployment(appId, branchName, jobId, sourceUrl)
}
\arguments{
\item{appId}{[required] The unique ID for an Amplify app.}

\item{branchName}{[required] The name for the branch, for the job.}

\item{jobId}{The job ID for this deployment, generated by the create deployment
request.}

\item{sourceUrl}{The source URL for this deployment, used when calling start deployment
without create deployment. The source URL can be any HTTP GET URL that
is publicly accessible and downloads a single .zip file.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  jobSummary = list(
    jobArn = "string",
    jobId = "string",
    commitId = "string",
    commitMessage = "string",
    commitTime = as.POSIXct(
      "2015-01-01"
    ),
    startTime = as.POSIXct(
      "2015-01-01"
    ),
    status = "PENDING"|"PROVISIONING"|"RUNNING"|"FAILED"|"SUCCEED"|"CANCELLING"|"CANCELLED",
    endTime = as.POSIXct(
      "2015-01-01"
    ),
    jobType = "RELEASE"|"RETRY"|"MANUAL"|"WEB_HOOK"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Starts a deployment for a manually deployed app. Manually deployed apps
are not connected to a repository.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$start_deployment(
  appId = "string",
  branchName = "string",
  jobId = "string",
  sourceUrl = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
