% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_start_job}
\alias{amplify_start_job}
\title{Starts a new job for a branch, part of an Amplify App}
\usage{
amplify_start_job(appId, branchName, jobId, jobType, jobReason,
  commitId, commitMessage, commitTime)
}
\arguments{
\item{appId}{[required] Unique Id for an Amplify App.}

\item{branchName}{[required] Name for the branch, for the Job.}

\item{jobId}{Unique Id for an existing job. Required for "RETRY" JobType.}

\item{jobType}{[required] Type for the Job. Available JobTypes are: \code{\\n} "RELEASE": Start a new
job with the latest change from the specified branch. Only available for
apps that have connected to a repository. "RETRY": Retry an existing
job. JobId is required for this type of job.}

\item{jobReason}{Descriptive reason for starting this job.}

\item{commitId}{Commit Id from 3rd party repository provider for the Job.}

\item{commitMessage}{Commit message from 3rd party repository provider for the Job.}

\item{commitTime}{Commit date / time for the Job.}
}
\description{
Starts a new job for a branch, part of an Amplify App.
}
\section{Request syntax}{
\preformatted{svc$start_job(
  appId = "string",
  branchName = "string",
  jobId = "string",
  jobType = "RELEASE"|"RETRY"|"MANUAL"|"WEB_HOOK",
  jobReason = "string",
  commitId = "string",
  commitMessage = "string",
  commitTime = as.POSIXct(
    "2015-01-01"
  )
)
}
}

\keyword{internal}
