% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_stop_job}
\alias{amplify_stop_job}
\title{Stop a job that is in progress, for an Amplify branch, part of Amplify
App}
\usage{
amplify_stop_job(appId, branchName, jobId)
}
\arguments{
\item{appId}{[required] Unique Id for an Amplify App.}

\item{branchName}{[required] Name for the branch, for the Job.}

\item{jobId}{[required] Unique Id for the Job.}
}
\description{
Stop a job that is in progress, for an Amplify branch, part of Amplify
App.
}
\section{Request syntax}{
\preformatted{svc$stop_job(
  appId = "string",
  branchName = "string",
  jobId = "string"
)
}
}

\keyword{internal}
