% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_update_domain_association}
\alias{amplify_update_domain_association}
\title{Creates a new domain association for an Amplify app}
\usage{
amplify_update_domain_association(appId, domainName,
  enableAutoSubDomain, subDomainSettings, autoSubDomainCreationPatterns,
  autoSubDomainIAMRole)
}
\arguments{
\item{appId}{[required] The unique ID for an Amplify app.}

\item{domainName}{[required] The name of the domain.}

\item{enableAutoSubDomain}{Enables the automated creation of subdomains for branches.}

\item{subDomainSettings}{[required] Describes the settings for the subdomain.}

\item{autoSubDomainCreationPatterns}{Sets the branch patterns for automatic subdomain creation.}

\item{autoSubDomainIAMRole}{The required AWS Identity and Access Management (IAM) service role for
the Amazon Resource Name (ARN) for automatically creating subdomains.}
}
\description{
Creates a new domain association for an Amplify app.
}
\section{Request syntax}{
\preformatted{svc$update_domain_association(
  appId = "string",
  domainName = "string",
  enableAutoSubDomain = TRUE|FALSE,
  subDomainSettings = list(
    list(
      prefix = "string",
      branchName = "string"
    )
  ),
  autoSubDomainCreationPatterns = list(
    "string"
  ),
  autoSubDomainIAMRole = "string"
)
}
}

\keyword{internal}
