% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_update_webhook}
\alias{amplify_update_webhook}
\title{Updates a webhook}
\usage{
amplify_update_webhook(webhookId, branchName, description)
}
\arguments{
\item{webhookId}{[required] The unique ID for a webhook.}

\item{branchName}{The name for a branch that is part of an Amplify app.}

\item{description}{The description for a webhook.}
}
\description{
Updates a webhook.
}
\section{Request syntax}{
\preformatted{svc$update_webhook(
  webhookId = "string",
  branchName = "string",
  description = "string"
)
}
}

\keyword{internal}
