% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsync_operations.R
\name{appsync_list_resolvers}
\alias{appsync_list_resolvers}
\title{Lists the resolvers for a given API and type}
\usage{
appsync_list_resolvers(apiId, typeName, nextToken, maxResults)
}
\arguments{
\item{apiId}{[required] The API ID.}

\item{typeName}{[required] The type name.}

\item{nextToken}{An identifier that was returned from the previous call to this
operation, which can be used to return the next set of items in the
list.}

\item{maxResults}{The maximum number of results you want the request to return.}
}
\description{
Lists the resolvers for a given API and type.
}
\section{Request syntax}{
\preformatted{svc$list_resolvers(
  apiId = "string",
  typeName = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
