% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsync_operations.R
\name{appsync_list_tags_for_resource}
\alias{appsync_list_tags_for_resource}
\title{Lists the tags for a resource}
\usage{
appsync_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The \code{GraphqlApi} ARN.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  tags = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Lists the tags for a resource.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_tags_for_resource(
  resourceArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
