% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_create_upload}
\alias{devicefarm_create_upload}
\title{Uploads an app or test scripts}
\usage{
devicefarm_create_upload(projectArn, name, type, contentType)
}
\arguments{
\item{projectArn}{[required] The ARN of the project for the upload.}

\item{name}{[required] The upload\'s file name. The name should not contain any forward slashes
(\code{/}). If you are uploading an iOS app, the file name must end with the
\code{.ipa} extension. If you are uploading an Android app, the file name
must end with the \code{.apk} extension. For all others, the file name must
end with the \code{.zip} file extension.}

\item{type}{[required] The upload\'s upload type.

Must be one of the following values:
\itemize{
\item ANDROID\\_APP
\item IOS\\_APP
\item WEB\\_APP
\item EXTERNAL\\_DATA
\item APPIUM\\_JAVA\\_JUNIT\\_TEST\\_PACKAGE
\item APPIUM\\_JAVA\\_TESTNG\\_TEST\\_PACKAGE
\item APPIUM\\_PYTHON\\_TEST\\_PACKAGE
\item APPIUM\\_NODE\\_TEST\\_PACKAGE
\item APPIUM\\_RUBY\\_TEST\\_PACKAGE
\item APPIUM\\_WEB\\_JAVA\\_JUNIT\\_TEST\\_PACKAGE
\item APPIUM\\_WEB\\_JAVA\\_TESTNG\\_TEST\\_PACKAGE
\item APPIUM\\_WEB\\_PYTHON\\_TEST\\_PACKAGE
\item APPIUM\\_WEB\\_NODE\\_TEST\\_PACKAGE
\item APPIUM\\_WEB\\_RUBY\\_TEST\\_PACKAGE
\item CALABASH\\_TEST\\_PACKAGE
\item INSTRUMENTATION\\_TEST\\_PACKAGE
\item UIAUTOMATION\\_TEST\\_PACKAGE
\item UIAUTOMATOR\\_TEST\\_PACKAGE
\item XCTEST\\_TEST\\_PACKAGE
\item XCTEST\\_UI\\_TEST\\_PACKAGE
\item APPIUM\\_JAVA\\_JUNIT\\_TEST\\_SPEC
\item APPIUM\\_JAVA\\_TESTNG\\_TEST\\_SPEC
\item APPIUM\\_PYTHON\\_TEST\\_SPEC
\item APPIUM\\_NODE\\_TEST\\_SPEC
\item APPIUM\\_RUBY\\_TEST\\_SPEC
\item APPIUM\\_WEB\\_JAVA\\_JUNIT\\_TEST\\_SPEC
\item APPIUM\\_WEB\\_JAVA\\_TESTNG\\_TEST\\_SPEC
\item APPIUM\\_WEB\\_PYTHON\\_TEST\\_SPEC
\item APPIUM\\_WEB\\_NODE\\_TEST\\_SPEC
\item APPIUM\\_WEB\\_RUBY\\_TEST\\_SPEC
\item INSTRUMENTATION\\_TEST\\_SPEC
\item XCTEST\\_UI\\_TEST\\_SPEC
}

If you call \code{CreateUpload} with \code{WEB_APP} specified, AWS Device Farm
throws an \code{ArgumentException} error.}

\item{contentType}{The upload\'s content type (for example, \code{application/octet-stream}).}
}
\description{
Uploads an app or test scripts.
}
\section{Request syntax}{
\preformatted{svc$create_upload(
  projectArn = "string",
  name = "string",
  type = "ANDROID_APP"|"IOS_APP"|"WEB_APP"|"EXTERNAL_DATA"|"APPIUM_JAVA_JUNIT_TEST_PACKAGE"|"APPIUM_JAVA_TESTNG_TEST_PACKAGE"|"APPIUM_PYTHON_TEST_PACKAGE"|"APPIUM_NODE_TEST_PACKAGE"|"APPIUM_RUBY_TEST_PACKAGE"|"APPIUM_WEB_JAVA_JUNIT_TEST_PACKAGE"|"APPIUM_WEB_JAVA_TESTNG_TEST_PACKAGE"|"APPIUM_WEB_PYTHON_TEST_PACKAGE"|"APPIUM_WEB_NODE_TEST_PACKAGE"|"APPIUM_WEB_RUBY_TEST_PACKAGE"|"CALABASH_TEST_PACKAGE"|"INSTRUMENTATION_TEST_PACKAGE"|"UIAUTOMATION_TEST_PACKAGE"|"UIAUTOMATOR_TEST_PACKAGE"|"XCTEST_TEST_PACKAGE"|"XCTEST_UI_TEST_PACKAGE"|"APPIUM_JAVA_JUNIT_TEST_SPEC"|"APPIUM_JAVA_TESTNG_TEST_SPEC"|"APPIUM_PYTHON_TEST_SPEC"|"APPIUM_NODE_TEST_SPEC"|"APPIUM_RUBY_TEST_SPEC"|"APPIUM_WEB_JAVA_JUNIT_TEST_SPEC"|"APPIUM_WEB_JAVA_TESTNG_TEST_SPEC"|"APPIUM_WEB_PYTHON_TEST_SPEC"|"APPIUM_WEB_NODE_TEST_SPEC"|"APPIUM_WEB_RUBY_TEST_SPEC"|"INSTRUMENTATION_TEST_SPEC"|"XCTEST_UI_TEST_SPEC",
  contentType = "string"
)
}
}

\examples{
# The following example creates a new Appium Python test package upload
# inside an existing project.
\donttest{svc$create_upload(
  name = "MyAppiumPythonUpload",
  type = "APPIUM_PYTHON_TEST_PACKAGE",
  projectArn = "arn:aws:devicefarm:us-west-2:123456789101:project:EXAMPLE-GUID-123-456"
)}

}
\keyword{internal}
