% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_install_to_remote_access_session}
\alias{devicefarm_install_to_remote_access_session}
\title{Installs an application to the device in a remote access session}
\usage{
devicefarm_install_to_remote_access_session(remoteAccessSessionArn,
  appArn)
}
\arguments{
\item{remoteAccessSessionArn}{[required] The Amazon Resource Name (ARN) of the remote access session about which
you are requesting information.}

\item{appArn}{[required] The ARN of the app about which you are requesting information.}
}
\description{
Installs an application to the device in a remote access session. For
Android applications, the file must be in .apk format. For iOS
applications, the file must be in .ipa format.
}
\section{Request syntax}{
\preformatted{svc$install_to_remote_access_session(
  remoteAccessSessionArn = "string",
  appArn = "string"
)
}
}

\examples{
# The following example installs a specific app to a device in a specific
# remote access session.
\dontrun{svc$install_to_remote_access_session(
  appArn = "arn:aws:devicefarm:us-west-2:123456789101:app:EXAMPLE-GUID-123-456",
  remoteAccessSessionArn = "arn:aws:devicefarm:us-west-2:123456789101:session:EXAMPLE-GUID-1..."
)}

}
\keyword{internal}
