% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_list_tags_for_resource}
\alias{devicefarm_list_tags_for_resource}
\title{List the tags for an AWS Device Farm resource}
\usage{
devicefarm_list_tags_for_resource(ResourceARN)
}
\arguments{
\item{ResourceARN}{[required] The Amazon Resource Name (ARN) of the resource(s) for which to list
tags. You can associate tags with the following Device Farm resources:
\code{PROJECT}, \code{RUN}, \code{NETWORK_PROFILE}, \code{INSTANCE_PROFILE},
\code{DEVICE_INSTANCE}, \code{SESSION}, \code{DEVICE_POOL}, \code{DEVICE}, and
\code{VPCE_CONFIGURATION}.}
}
\description{
List the tags for an AWS Device Farm resource.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  ResourceARN = "string"
)
}
}

\keyword{internal}
