% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_list_test_grid_projects}
\alias{devicefarm_list_test_grid_projects}
\title{Gets a list of all Selenium testing projects in your account}
\usage{
devicefarm_list_test_grid_projects(maxResult, nextToken)
}
\arguments{
\item{maxResult}{Return no more than this number of results.}

\item{nextToken}{From a response, used to continue a paginated listing.}
}
\description{
Gets a list of all Selenium testing projects in your account.
}
\section{Request syntax}{
\preformatted{svc$list_test_grid_projects(
  maxResult = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
