% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_list_test_grid_session_artifacts}
\alias{devicefarm_list_test_grid_session_artifacts}
\title{Retrieves a list of artifacts created during the session}
\usage{
devicefarm_list_test_grid_session_artifacts(sessionArn, type, maxResult,
  nextToken)
}
\arguments{
\item{sessionArn}{[required] The ARN of a TestGridSession.}

\item{type}{Limit results to a specified type of artifact.}

\item{maxResult}{The maximum number of results to be returned by a request.}

\item{nextToken}{Pagination token.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  artifacts = list(
    list(
      filename = "string",
      type = "UNKNOWN"|"VIDEO"|"SELENIUM_LOG",
      url = "string"
    )
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves a list of artifacts created during the session.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_test_grid_session_artifacts(
  sessionArn = "string",
  type = "VIDEO"|"LOG",
  maxResult = 123,
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
