% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_renew_offering}
\alias{devicefarm_renew_offering}
\title{Explicitly sets the quantity of devices to renew for an offering,
starting from the effectiveDate of the next period}
\usage{
devicefarm_renew_offering(offeringId, quantity)
}
\arguments{
\item{offeringId}{The ID of a request to renew an offering.}

\item{quantity}{The quantity requested in an offering renewal.}
}
\description{
Explicitly sets the quantity of devices to renew for an offering,
starting from the \code{effectiveDate} of the next period. The API returns a
\code{NotEligible} error if the user is not permitted to invoke the
operation. If you must be able to invoke this operation, contact
\href{mailto:aws-devicefarm-support@amazon.com}{aws-devicefarm-support@amazon.com}.
}
\section{Request syntax}{
\preformatted{svc$renew_offering(
  offeringId = "string",
  quantity = 123
)
}
}

\examples{
\dontrun{
# The following example renews a specific device slot offering.
svc$renew_offering(
  offeringId = "D68B3C05-1BA6-4360-BC69-12345EXAMPLE",
  quantity = 1L
)
}

}
\keyword{internal}
