% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_update_test_grid_project}
\alias{devicefarm_update_test_grid_project}
\title{Change details of a project}
\usage{
devicefarm_update_test_grid_project(projectArn, name, description)
}
\arguments{
\item{projectArn}{[required] ARN of the project to update.}

\item{name}{Human-readable name for the project.}

\item{description}{Human-readable description for the project.}
}
\value{
A list with the following syntax:\preformatted{list(
  testGridProject = list(
    arn = "string",
    name = "string",
    description = "string",
    created = as.POSIXct(
      "2015-01-01"
    )
  )
)
}
}
\description{
Change details of a project.
}
\section{Request syntax}{
\preformatted{svc$update_test_grid_project(
  projectArn = "string",
  name = "string",
  description = "string"
)
}
}

\keyword{internal}
