% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobileanalytics_service.R
\name{mobileanalytics}
\alias{mobileanalytics}
\title{Amazon Mobile Analytics}
\usage{
mobileanalytics(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
Amazon Mobile Analytics is a service for collecting, visualizing, and
understanding app usage data at scale.
}
\section{Service syntax}{
\preformatted{svc <- mobileanalytics(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=mobileanalytics_put_events]{put_events} \tab The PutEvents operation records one or more events
}
}

\examples{
\dontrun{svc <- mobileanalytics()
svc$put_events(
  Foo = 123
)}

}
