% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mobile_operations.R
\name{update_project}
\alias{update_project}
\title{Update an existing project}
\usage{
update_project(contents = NULL, projectId)
}
\arguments{
\item{contents}{ZIP or YAML file which contains project configuration to be updated. This should be the contents of the file downloaded from the URL provided in an export project operation.}

\item{projectId}{[required] Unique project identifier.}
}
\description{
Update an existing project.
}
\section{Accepted Parameters}{
\preformatted{update_project(
  contents = raw,
  projectId = "string"
)
}
}

